% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{addCohortPermission}
\alias{addCohortPermission}
\title{Assign cohort definition permissions to user.}
\usage{
addCohortPermission(baseUrl, cohortId, userId, permission)
}
\arguments{
\item{baseUrl}{The URL to WebAPI}

\item{cohortId}{The ID of the cohort definition}

\item{userId}{The userId to grant permission}

\item{permission}{The permission type to assign.}
}
\description{
Assign cohort definition permissions to user.
}
\details{
The function will call the webservice to assign permissions to a cohort definition for the given userId.
}
\examples{
\dontrun{
getCohortSql(CohortDefinition = (getCohortDefinition(cohortId = 13242, baseUrl = baseUrl)),
             baseUrl = "http://server.org:80/WebAPI")
}
}
