% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generation.R
\name{cancelGeneration}
\alias{cancelGeneration}
\title{Invoke generation.}
\usage{
cancelGeneration(id, baseUrl, sourceKey, category)
}
\arguments{
\item{id}{An integer id representing the id that uniquely identifies a 
definition for the category in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}

\item{category}{These are the categories in WebApi. The valid string options are 'conceptSet', 
'cohort', 'characterization', 'pathway, 'incidenceRate', 'estimation', 'prediction'.}
}
\value{
Error message if invoke failed.
}
\description{
Invoke generation.
}
\details{
Invoke generation (execution) information.
}
\examples{
\dontrun{
cancelGeneration(id = 13242, category = "cohort", baseUrl = "http://server.org:80/WebAPI")
}
}
