% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{cancelIncidenceRateGeneration}
\alias{cancelIncidenceRateGeneration}
\title{Cancel generation of IncidenceRate id.}
\usage{
cancelIncidenceRateGeneration(incidenceRateId, baseUrl, sourceKey)
}
\arguments{
\item{incidenceRateId}{An integer id representing the id that uniquely identifies a IncidenceRate
definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}
}
\value{
A tibble with job status information.
}
\description{
Cancel generation of IncidenceRate id.
}
\details{
Cancel the generation of IncidenceRate id in the WebApi.
}
\examples{
\dontrun{
cancelIncidenceRateGeneration(incidenceRateId = 13242,
                              baseUrl = "http://server.org:80/WebAPI",
                              sourceKey = "HCUP")
}
}
