% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{checkIfEstimationNameExists}
\alias{checkIfEstimationNameExists}
\title{Check if Estimation definition name exists.}
\usage{
checkIfEstimationNameExists(estimationName, baseUrl)
}
\arguments{
\item{estimationName}{A string name for the Estimation to be checked.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
If found, the function will return a tibble with details of the specification. If not found, FALSE
will be returned.
}
\description{
Check if Estimation definition name exists.
}
\details{
Check if a string name already exists in the WebApi as a Estimation definition name.
}
\examples{
\dontrun{
checkIfEstimationNameExists(estimationName = "this text string needs to be checked",
                            baseUrl = "http://server.org:80/WebAPI")
}
}
