% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auth.R
\name{connectWebApi}
\alias{connectWebApi}
\title{Create a connection to an instance of WebAPI}
\usage{
connectWebApi(
  baseUrl,
  authMethod = guessAuthMethod(),
  atlasUsername = NULL,
  keyringServiceName = NULL
)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{authMethod}{The method used for authentication to WebAPI. Options are
\itemize{
 \item{"db"}{Database authentication using Atlas/WebAPI built in auth}
 \item{"AD"}{Active Directory}
 \item{"windows"}{Windows}
 \item{"none"}{No authentication. WebAPI security must be disabled.}
}
By default the correct authentication method will be guessed.}

\item{keyringServiceName}{(Optional) Keyring stores passwords securely. 
For some authentication methods it may be convenient 
to pass keyring service name instead of entering a password.}
}
\value{
A WebApi connection object
}
\description{
This function creates a WebAPI connection object that can be used by other
functions in this package to access WebAPI. This function will test 
the connection as well as trigger authentication and store the auth
session token.
}
