% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{convertConceptSetDefinitionToTable}
\alias{convertConceptSetDefinitionToTable}
\title{Convert a concept set definition to a table}
\usage{
convertConceptSetDefinitionToTable(
  conceptSetDefinition,
  snakeCaseToCamelCase = TRUE
)
}
\arguments{
\item{conceptSetDefinition}{A concept set definition, for example as obtained through the \code{\link{getConceptSetDefinition}} function, 
or taken from a cohort definition.}

\item{snakeCaseToCamelCase}{Should the column names of the result be converted to camelCase?}
}
\value{
A tibble representing the concept set expression.
}
\description{
Convert a concept set definition to a table
}
\examples{
\dontrun{
conceptSetDefinition <- getConceptSetDefinition(conceptSetId = 282, 
                                                baseUrl = "http://server.org:80/WebAPI")
convertConceptSetDefinitionToTable(conceptSetDefinition = conceptSetDefinition)
}

}
