% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{convertConceptSetToTable}
\alias{convertConceptSetToTable}
\title{Convert a concept set to a table}
\usage{
convertConceptSetToTable(conceptSet, snakeCaseToCamelCase = TRUE)
}
\arguments{
\item{conceptSet}{A concept set, for example as obtained through the \code{\link{getConceptSet}} function, 
or taken from a cohort definition.}

\item{snakeCaseToCamelCase}{Should the column names of the result be converted to camelCase?}
}
\value{
A tibble representing the concept set expression.
}
\description{
Convert a concept set to a table
}
\examples{
\dontrun{
conceptSet <- getConceptSet(conceptSetId = 282, baseUrl = "http://server.org:80/WebAPI")
conceptIds <- resolveConceptSet(conceptSet = conceptSet,
                                baseUrl = "http://server.org:80/WebAPI")
}

}
