% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{deleteCohortDefinition}
\alias{deleteCohortDefinition}
\title{Delete a cohort definition}
\usage{
deleteCohortDefinition(cohortId, baseUrl, silent = FALSE, stopOnError = FALSE)
}
\arguments{
\item{cohortId}{The number indicating which cohort definition to fetch.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{silent}{[OPTIONAL, Default = FALSE] If TRUE, function will work silently without showing any warning or error message.}

\item{stopOnError}{[OPTIONAL, Default = FALSE] If silent silent = TRUE, then this will be ignored.}
}
\value{
NA. A status message will be shown.
}
\description{
Delete a cohort definition
}
\details{
Deletes cohort definition from WebAPI for a given cohort id
}
\examples{
\dontrun{
deleteCohortDefinition(cohortId = 282, baseUrl = "http://server.org:80/WebAPI")
}

}
