% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{deleteConceptSetDefinition}
\alias{deleteConceptSetDefinition}
\title{Delete ConceptSet id definition.}
\usage{
deleteConceptSetDefinition(conceptSetId, baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{ConceptSetId}{An integer id representing the id that uniquely identifies a ConceptSet
definition in a WebApi instance.}
}
\value{
An R object representing the ConceptSet definition
}
\description{
Delete ConceptSet id definition.
}
\details{
Delete the ConceptSet definition from WebAPI for a given ConceptSet id
}
\examples{
\dontrun{
getConceptSetDefinition(ConceptSetId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
