% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{deleteEstimationDefinition}
\alias{deleteEstimationDefinition}
\title{Delete Estimation id definition.}
\usage{
deleteEstimationDefinition(estimationId, baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{EstimationId}{An integer id representing the id that uniquely identifies a Estimation
definition in a WebApi instance.}
}
\value{
An R object representing the Estimation definition
}
\description{
Delete Estimation id definition.
}
\details{
Delete the Estimation definition from WebAPI for a given Estimation id
}
\examples{
\dontrun{
getEstimationDefinition(EstimationId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
