% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{deletePredictionDefinition}
\alias{deletePredictionDefinition}
\title{Delete Prediction id definition.}
\usage{
deletePredictionDefinition(predictionId, baseUrl)
}
\arguments{
\item{predictionId}{An integer id representing the id that uniquely identifies a Prediction
definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
An R object representing the Prediction definition
}
\description{
Delete Prediction id definition.
}
\details{
Delete the Prediction definition from WebAPI for a given Prediction id
}
\examples{
\dontrun{
deletePredictionDefinition(predictionId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
