% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{detectCharacterizationsByName}
\alias{detectCharacterizationsByName}
\title{Detect the presence of string matched Characterization definitions.}
\usage{
detectCharacterizationsByName(pattern, negate = FALSE, baseUrl)
}
\arguments{
\item{pattern}{A pattern to look for. See \link[stringr]{str_detect} for details.}

\item{negate}{If TRUE, return non-matching elements. See \link[stringr]{str_detect} for details.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
FALSE if no matches. If matched - output from
\link[ROhdsiWebApi]{getCharacterizationDefinitionsMetaData}
}
\description{
Detect the presence of string matched Characterization definitions.
}
\details{
Detect string matched Characterization definition names from the WebApi, and retrieve metadata
definitions.
}
\examples{
\dontrun{
detectCharacterizations(pattern = "this text string to search in pattern",
                        baseUrl = "http://server.org:80/WebAPI")
}
}
