% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{detectEstimationsByName}
\alias{detectEstimationsByName}
\title{Detect the presence of string matched Estimation definitions.
\lifecycle{stable}}
\usage{
detectEstimationsByName(pattern, negate = FALSE, baseUrl)
}
\arguments{
\item{pattern}{A pattern to look for. See \link[stringr]{str_detect} for details.}

\item{negate}{If TRUE, return non-matching elements. See \link[stringr]{str_detect} for details.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
FALSE if no matches. If matched - output from \link[ROhdsiWebApi]{getEstimationDefinitionsMetaData}
}
\description{
Detect the presence of string matched Estimation definitions.
\lifecycle{stable}
}
\details{
Detect string matched Estimation definition names from the WebApi, and retrieve
metadata definitions.
}
\examples{
\dontrun{
detectEstimations(pattern = 'this text string to search in pattern', 
baseUrl = "http://server.org:80/WebAPI")
}
}
