% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auth.R
\name{.authDb}
\alias{.authDb}
\title{Authenticate using Atlas database authentication}
\usage{
.authDb(baseUrl, webApiUsername, webApiPassword)
}
\value{
a bearer token to be used in http headers in calls to WebAPI
}
\description{
Authenticate using Atlas database authentication
}
