% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{existsConceptSetName}
\alias{existsConceptSetName}
\title{Check if ConceptSet definition name exists.}
\usage{
existsConceptSetName(conceptSetName, baseUrl)
}
\arguments{
\item{conceptSetName}{A string name for the ConceptSet to be checked.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
If found, the function will return a tibble with details of the specification. If not found, FALSE
will be returned.
}
\description{
Check if ConceptSet definition name exists.
}
\details{
Check if a string name already exists in the WebApi as a ConceptSet definition name.
}
\examples{
\dontrun{
existsConceptSetName(conceptSetName = "this text string needs to be checked",
                     baseUrl = "http://server.org:80/WebAPI")
}
}
