% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{existsEstimationName}
\alias{existsEstimationName}
\title{Check if Estimation definition name exists.
\lifecycle{stable}}
\usage{
existsEstimationName(estimationName, baseUrl)
}
\arguments{
\item{estimationName}{A string name for the Estimation to be checked.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
If found, the function will return a tibble with details of the specification.
                         If not found, FALSE will be returned.
}
\description{
Check if Estimation definition name exists.
\lifecycle{stable}
}
\details{
Check if a string name already exists in the WebApi as a Estimation definition name.
}
\examples{
\dontrun{
existsEstimationName(estimationName = 'this text string needs to be checked', 
baseUrl = "http://server.org:80/WebAPI")
}
}
