% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCohortDefinitionSet.R
\name{exportCohortDefinitionSet}
\alias{exportCohortDefinitionSet}
\title{Export cohort definition set from WebAPI}
\usage{
exportCohortDefinitionSet(baseUrl, cohortIds = c(), generateStats)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{cohortIds}{A set of cohortIds to fetch from WebApi.}

\item{generateStats}{Should cohort inclusion rule statistics be generated?}
}
\description{
Export cohort definition set from WebAPI
}
\details{
Constructs a CohortDefinition set containing the following fields: \describe{
\item{atlasId}{The cohort ID in
ATLAS.} \item{cohortId}{a copy of the value in atlasId.} \item{cohortName}{The name of the cohort.} 
\item{sql}{The cohort generation sql.} \item{json}{The cohort definition JSON.}
\item{logicDescription}{The cohort description.}}
}
