% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{formatDfName}
\alias{formatDfName}
\title{Formats name by removing prefix. If the prefix starts with '[' and ends with ']'}
\usage{
formatDfName(name)
}
\arguments{
\item{name}{Name needing formatting}
}
\value{
renamed name
}
\description{
Formats name by removing prefix. If the prefix starts with '[' and ends with ']'
}
\details{
Formats name by removing prefix. If the prefix starts with '[' and ends with ']'
}
