% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httrWrappers.R
\name{GET}
\alias{GET}
\title{Internal wrappers for httr request functions}
\usage{
GET(url, config = list(), ..., handle = NULL)
}
\description{
The primary reason for this wrapper is to add the appropriate authorization header to requests.
This file also centralizes http error handling.
}
