% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{getBearerDbLogin}
\alias{getBearerDbLogin}
\title{Get the data sources in the WebAPI instance}
\usage{
getBearerDbLogin(baseUrl, userLogin, userPassword)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{userLogin}{Login}

\item{userPassword}{Password}
}
\value{
A data frame of data source information
}
\description{
Get the data sources in the WebAPI instance
}
\details{
Obtains the data sources configured in the WebAPI instance
}
