% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{getCdmSources}
\alias{getCdmSources}
\title{Get the data sources in the WebAPI instance \lifecycle{stable}}
\usage{
getCdmSources(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A data frame.
}
\description{
Get the data sources in the WebAPI instance \lifecycle{stable}
}
\details{
Obtains the data sources configured in the WebAPI instance.
}
