% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getCharacterizationDefinition}
\alias{getCharacterizationDefinition}
\title{Get Characterization id definition. \lifecycle{stable}}
\usage{
getCharacterizationDefinition(characterizationId, baseUrl)
}
\arguments{
\item{characterizationId}{An integer id representing the id that uniquely identifies a
Characterization definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
An R object representing the Characterization definition
}
\description{
Get Characterization id definition. \lifecycle{stable}
}
\details{
Obtain the Characterization definition from WebAPI for a given Characterization id
}
\examples{
\dontrun{
getCharacterizationDefinition(characterizationId = 13242,
                              baseUrl = "http://server.org:80/WebAPI")
}
}
