% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\name{getCharacterizationDefinitionExpression}
\alias{getCharacterizationDefinitionExpression}
\title{Get a characterization definition expression}
\usage{
getCharacterizationDefinitionExpression(
  baseUrl,
  characterizationId,
  generationId = NULL
)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{characterizationId}{The number indicating which characterization definition to fetch.}

\item{generationId}{(OPTIONAL) Used to specify the id of a particular generation of a cohort
characterization. If generationId is provided, then characterizationId is ignored.}
}
\value{
A JSON object representing the characterizationId/generationId definition returned by webApi.
A warning will be shown if the characterizationId/generationId does not exist.
}
\description{
Get a characterization definition expression
}
\details{
Obtain the characterization definition expression from WebAPI for a given characterization id/generation id
}
\examples{
\dontrun{
getCharacterizationDefinitionExpression(characterizationId = 282, baseUrl = "http://server.org:80/WebAPI")
}

}
