% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getCharacterizationDefinitionsMetaData}
\alias{getCharacterizationDefinitionsMetaData}
\title{Get the meta data for Characterization definitions.}
\usage{
getCharacterizationDefinitionsMetaData(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
A tibble of specification metadata for Characterization. Note: modifiedDate and createdDate are
returned as text/character.
}
\description{
Get the meta data for Characterization definitions.
}
\details{
Get the meta data of WebApi specifications such as id, name, created/modified details, hash object,
etc. from WebApi for Characterization. This function is useful to retrieve the current
Characterization specifications.
}
\examples{
\dontrun{
getCharacterizationDefinitionsMetaData(baseUrl = "http://server.org:80/WebAPI")
}
}
