% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\name{getCohortCharacterizationDefinition}
\alias{getCohortCharacterizationDefinition}
\title{Get a characterization definition expression}
\usage{
getCohortCharacterizationDefinition(
  characterizationId,
  baseUrl,
  generationId = NULL
)
}
\arguments{
\item{characterizationId}{The number indicating which characterization definition to fetch.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{generationId}{(OPTIONAL) Used to specify the id of a particular generation of a cohort
characterization. If generationId is provided, then characterizationId is ignored.}
}
\value{
A R-object representing the characterizationId/generationId definition returned by WebApi.
A warning will be shown if the characterizationId/generationId does not exist.
}
\description{
Get a characterization definition expression
}
\details{
Obtain the characterization definition expression from WebAPI as R-object for a given characterization id/generation id
}
\examples{
\dontrun{
getCohortCharacterizationDefinition(characterizationId = 282, 
                                    baseUrl = "http://server.org:80/WebAPI")
}

}
