% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\name{getCohortCharacterizationResults}
\alias{getCohortCharacterizationResults}
\title{Get Cohort Characterization Results}
\usage{
getCohortCharacterizationResults(baseUrl, characterizationId,
  generationId = NULL, sourceKey, domains = c(), analysisNames = c())
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{characterizationId}{The id of the cohort characterization in Atlas}

\item{generationId}{Used to specify the id of a particular generation of a cohort characterization.
By default, the latest execution is retrieved}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the Configuration page}

\item{domains}{(OPTIONAL) Which feature domains would you like to retrieve?}

\item{analysisNames}{(OPTIONAL) Which feature analysis names would you like to retrieve?}
}
\description{
Get Cohort Characterization Results
}
