% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\name{getCohortCharacterizationSpecification}
\alias{getCohortCharacterizationSpecification}
\title{Get Cohort Characterization Specification}
\usage{
getCohortCharacterizationSpecification(
  baseUrl,
  characterizationId,
  generationId = NULL
)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{characterizationId}{The id of the cohort characterization in Atlas}

\item{generationId}{(OPTIONAL) Used to specify the id of a particular generation of a cohort
characterization. By default, the latest execution is retrieved}
}
\description{
Get Cohort Characterization Specification
}
