% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getCohortDefinition}
\alias{getCohortDefinition}
\title{Get a cohort definition}
\usage{
getCohortDefinition(cohortId, baseUrl)
}
\arguments{
\item{cohortId}{The number indicating which cohort definition to fetch.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
An R object representing the cohort definition
}
\description{
Get a cohort definition
}
\details{
Obtain the cohort definition from WebAPI for a given cohort id
}
\examples{
\dontrun{
getCohortDefinition(cohortId = 282, baseUrl = "http://server.org:80/WebAPI")
}

}
