% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getCohortDefinitionExpression}
\alias{getCohortDefinitionExpression}
\title{Get a cohort definition expression}
\usage{
getCohortDefinitionExpression(definitionId, baseUrl)
}
\arguments{
\item{definitionId}{The number indicating which cohort definition to fetch.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
A JSON list object representing the cohort definition
This function has been deprecated. As an alternative please use the following
steps as in the example below:
  1) validJsonExpression <- getCohortDefinition(baseUrl = baseUrl, cohortId = 15873)
  2) validJsonExpression <- RJSONIO::toJSON(cohortDefinition$expression)
  3) save validJsonExpression object as .txt"
}
\description{
Get a cohort definition expression
}
\details{
Obtain the JSON expression from WebAPI for a given cohort id
}
\examples{
\dontrun{
# This will obtain a cohort definition's JSON expression:

getCohortDefinitionExpression(definitionId = 282, baseUrl = "http://server.org:80/WebAPI")
}

}
