% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getCohortDefinitionName}
\alias{getCohortDefinitionName}
\title{Get a cohort definition's name from WebAPI}
\usage{
getCohortDefinitionName(baseUrl, definitionId, formatName = FALSE)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{definitionId}{The cohort definition id in Atlas.}

\item{formatName}{Should the name be formatted to remove prefixes and underscores?}
}
\value{
The name of the cohort.
}
\description{
Get a cohort definition's name from WebAPI
}
\details{
Obtains the name of a cohort.
}
