% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getCohortDefinitionsMetaData}
\alias{getCohortDefinitionsMetaData}
\title{Get the meta data for Cohort definitions.}
\usage{
getCohortDefinitionsMetaData(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
A tibble of specification metadata for Cohort. Note: modifiedDate and createdDate are returned as
text/character.
}
\description{
Get the meta data for Cohort definitions.
}
\details{
Get the meta data of WebApi specifications such as id, name, created/modified details, hash object,
etc. from WebApi for Cohort. This function is useful to retrieve the current Cohort specifications.
}
\examples{
\dontrun{
getCohortDefinitionsMetaData(baseUrl = "http://server.org:80/WebAPI")
}
}
