% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getCohortGenerationInformation}
\alias{getCohortGenerationInformation}
\title{Get cohort generation information}
\usage{
getCohortGenerationInformation(cohortIds, baseUrl, sourceKeys = NULL)
}
\arguments{
\item{cohortIds}{A list of cohortIds}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{sourceKeys}{(OPTIONAL) A list of CDM source keys. These can be found in Atlas ->
Configure. Otherwise, all CDM source keys will be used.}
}
\value{
A data frame of cohort generation statuses, start times, and execution durations per definition id
and source key.
}
\description{
Get cohort generation information
}
\details{
Obtains cohort generation statuses for a collection of cohort definition Ids and CDM sources.
Useful if running multiple cohort generation jobs that are long-running.
}
