% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getCohortGenerationReport}
\alias{getCohortGenerationReport}
\title{Get cohort generation report}
\usage{
getCohortGenerationReport(baseUrl, cohortId, sourceKey, mode = 0)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".}

\item{cohortId}{The Atlas cohort definition id for the cohort}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the Configuration page}

\item{mode}{Mode is used to differentiate between inclusion rules and count by events (mode = 0, default) 
or persons (mode = 1). Default value = 0.}
}
\value{
A list of data frames containing cohort generation report
}
\description{
Get cohort generation report
}
\details{
Obtains a list with dataframe containing details of report for cohort generation output.
}
