% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getCohortResults}
\alias{getCohortResults}
\title{Get cohort generation results}
\usage{
getCohortResults(cohortId, baseUrl, sourceKey, mode = 0)
}
\arguments{
\item{cohortId}{An integer id representing the id that uniquely identifies a cohort definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}

\item{mode}{Mode is used to differentiate between inclusion rules and count by events (mode = 0, default) 
or persons (mode = 1). Default value = 0.}
}
\value{
A list of data frames containing cohort generation report
}
\description{
Get cohort generation results
}
\details{
Obtains a list with data frame containing details of output for cohort generation
}
\examples{
\dontrun{
getCohortGeneratioInformation(cohortId = 282, 
                              baseUrl = "http://server.org:80/WebAPI", 
                              sourceKey = "HCUP", 
                              mode = 1)
}
}
