% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getCohortSql}
\alias{getCohortSql}
\title{Get SQL query for Cohort definition.}
\usage{
getCohortSql(cohortDefinition, baseUrl)
}
\arguments{
\item{cohortDefinition}{An R list object (not JSON) representing the Cohort definition. It is the
output R expression object of list object from \code{CohortDefinition}}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
An R object containing the SQL for Cohort definition.
}
\description{
Get SQL query for Cohort definition.
}
\details{
Given a valid Cohort definition R-object (not JSON) this function will return the parameterized SQL
in OHDSI SQL dialect. This SQL then may be used along with OHDSI R-package 'SQLRender' to
render/translate to target SQL dialect and parameters rendered.
}
\examples{
\dontrun{
getCohortSql(CohortDefinition = (getCohortDefinition(cohortId = 13242, baseUrl = baseUrl)),
             baseUrl = "http://server.org:80/WebAPI")
}
}
