% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concept.R
\name{getConcepts}
\alias{getConcepts}
\title{Get concepts \lifecycle{stable}}
\usage{
getConcepts(
  conceptIds,
  baseUrl,
  vocabularySourceKey = NULL,
  snakeCaseToCamelCase = TRUE
)
}
\arguments{
\item{conceptIds}{A vector of concept IDs.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{vocabularySourceKey}{The source key of the Vocabulary. By default, the priority Vocabulary is
used.}

\item{snakeCaseToCamelCase}{Should the column names of the result be converted to camelCase?}
}
\value{
A tibble containing information on the concepts.
}
\description{
Get concepts \lifecycle{stable}
}
\examples{
\dontrun{
conceptSet <- getConceptSet(conceptSetId = 282, baseUrl = "http://server.org:80/WebAPI")
conceptIds <- resolveConceptSet(conceptSet = conceptSet, baseUrl = "http://server.org:80/WebAPI")
concepts <- getConcepts(conceptIds = conceptIds, baseUrl = "http://server.org:80/WebAPI")
}

}
