% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getConceptSet}
\alias{getConceptSet}
\title{Get a concept set}
\usage{
getConceptSet(baseUrl, conceptSetId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{conceptSetId}{An integer id representing the id that uniquely identifies a 
concept set definition in a WebApi instance.}
}
\value{
An R object representing the concept set
}
\description{
Get a concept set
}
\details{
Obtain a concept set from WebAPI
}
\examples{
\dontrun{
conceptSet <- getConceptSet(conceptSetId = 282, baseUrl = "http://server.org:80/WebAPI")
}

}
