% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getConceptSetConceptIds}
\alias{getConceptSetConceptIds}
\title{Get Concept Set Concept Ids}
\usage{
getConceptSetConceptIds(baseUrl, conceptSetId, vocabSourceKey = NULL)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{conceptSetId}{An integer id representing the id that uniquely identifies a 
concept set definition in a WebApi instance.}

\item{vocabSourceKey}{The source key of the Vocabulary. By default, the priority Vocabulary is
used.}
}
\value{
A list of concept Ids.
}
\description{
Get Concept Set Concept Ids
}
\details{
Obtains the full list of concept Ids in a concept set.
}
