% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getConceptSetDefinition}
\alias{getConceptSetDefinition}
\title{Get a concept set definition}
\usage{
getConceptSetDefinition(conceptSetId, baseUrl)
}
\arguments{
\item{conceptSetId}{An integer id representing the id that uniquely identifies a 
concept set definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
An R object representing the concept set.
}
\description{
Get a concept set definition
}
\details{
Obtain a concept set from WebAPI.
}
\examples{
\dontrun{
conceptSetDefinition <- getConceptSetDefinition(conceptSetId = 282, 
                                                baseUrl = "http://server.org:80/WebAPI")
}

}
