% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConceptSetDefinitionBySourceKey.R
\name{getConceptSetDefinitionBySourceKey}
\alias{getConceptSetDefinitionBySourceKey}
\title{Fetch concept set definition from WebAPI by SourceKey}
\usage{
getConceptSetDefinitionBySourceKey(
  conceptSetId,
  baseUrl,
  vocabularySourceKey = NULL
)
}
\arguments{
\item{conceptSetId}{the id of the concept set to retrieve.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{vocabularySourceKey}{The source key of the Vocabulary. By default, the priority Vocabulary is
used.}
}
\value{
An R object representing the ConceptSet definition
}
\description{
Fetch concept set definition from WebAPI by SourceKey
}
\details{
Fetches a concept set definition from WebAPI by SourceKey. If SourceKey is not 
specified, the priority vocabulary will be used.
}
\examples{
\dontrun{
conceptSetDefinition <- getConceptSetDefinitionBySourceKey(conceptSetId = 282,
                                                baseUrl = "http://server.org:80/WebAPI",
                                                vocabularySourceKey = "MY_VOCAB")
conceptIds <- resolveConceptSet(conceptSetDefinition = conceptSetDefinition,
                                baseUrl = "http://server.org:80/WebAPI")
}

}
