% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getConceptSetDefinitionsMetaData}
\alias{getConceptSetDefinitionsMetaData}
\title{Get the meta data for ConceptSet definitions.}
\usage{
getConceptSetDefinitionsMetaData(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
A tibble of specification metadata for ConceptSet. Note: modifiedDate and createdDate are returned
as text/character.
}
\description{
Get the meta data for ConceptSet definitions.
}
\details{
Get the meta data of WebApi specifications such as id, name, created/modified details, hash object,
etc. from WebApi for ConceptSet. This function is useful to retrieve the current ConceptSet
specifications.
}
\examples{
\dontrun{
getConceptSetDefinitionsMetaData(baseUrl = "http://server.org:80/WebAPI")
}
}
