% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getConceptSetExpression}
\alias{getConceptSetExpression}
\title{Get a concept set expression}
\usage{
getConceptSetExpression(baseUrl, conceptSetId, asDataFrame = FALSE)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{conceptSetId}{An integer id representing the id that uniquely identifies a 
concept set definition in a WebApi instance.}

\item{asDataFrame}{(OPTIONAL) Get expression as data frame}
}
\value{
A JSON list object representing the concept set
}
\description{
Get a concept set expression
}
\details{
Obtain the JSON expression from WebAPI for a given concept set
}
\examples{
\dontrun{
# This will obtain a concept set's JSON expression:

getConceptSetExpression(conceptSetId = 282, baseUrl = "http://server.org:80/WebAPI")
}

}
