% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getConceptSetExpressionFromCohort}
\alias{getConceptSetExpressionFromCohort}
\title{Get a list of concept set expression from a cohort definition}
\usage{
getConceptSetExpressionFromCohort(baseUrl, definitionId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{definitionId}{The cohort id to fetch concept set expressions}
}
\value{
A list of concept set expressions
}
\description{
Get a list of concept set expression from a cohort definition
}
\details{
For a given cohort definition id, get all concept set expressions
}
\examples{
\dontrun{
# This will obtain a list of concept set expressions from a cohort id:

getConceptSetExpressionFromCohort(baseUrl = "http://server.org:80/WebAPI",
                        definitionId = 123)
}

}
