% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getConceptSetName}
\alias{getConceptSetName}
\title{Get a concept set's name from WebAPI}
\usage{
getConceptSetName(baseUrl, setId, formatName = FALSE)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{setId}{The concept set id in Atlas.}

\item{formatName}{Should the name be formatted to remove prefixes and underscores?}
}
\value{
The name of the concept set.
}
\description{
Get a concept set's name from WebAPI
}
\details{
Obtains the name of a concept set.
}
