% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getConceptSets}
\alias{getConceptSets}
\title{Get the available concept sets}
\usage{
getConceptSets(baseUrl, searchPrefix = "")
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{searchPrefix}{A search prefix to limit the concept sets by name}
}
\value{
A data frame of concept set names
}
\description{
Get the available concept sets
}
\details{
Obtains the concept sets
}
