% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getConceptSetsFromCohortDefinition}
\alias{getConceptSetsFromCohortDefinition}
\title{Get concept set expressions from cohort definition \lifecycle{maturing}}
\usage{
getConceptSetsFromCohortDefinition(cohortDefinition)
}
\arguments{
\item{cohortDefinition}{An R-object (not JSON) representing the cohort definition}
}
\value{
A tibble data frame with all the concept sets in the cohort definition R-object
}
\description{
Get all concept set expressions from a cohort definition object as a signel dataframe
}
