% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getConceptSetsSourceCodes}
\alias{getConceptSetsSourceCodes}
\title{Get a concept set's included source codes}
\usage{
getConceptSetsSourceCodes(
  baseUrl,
  setId,
  vocabSourceKey = NULL,
  vocabularyIds = c()
)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{setId}{The concept set id in Atlas.}

\item{vocabSourceKey}{The source key of the Vocabulary. By default, the priority Vocabulary is
used.}

\item{vocabularyIds}{An array of source code vocabulary Ids to filter the results by}
}
\value{
A data frame of included source codes
}
\description{
Get a concept set's included source codes
}
\details{
Obtains the included source codes from a concept set
}
