% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{getConfiguration}
\alias{getConfiguration}
\title{Get the WebApi configuration for the WebApi instance.}
\usage{
getConfiguration(baseUrl, sourceKeys)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceKeys}{A list of string names for the source keys as in
configuration page of Atlas.}
}
\value{
A list representing the WebApi configuration for the CDM Schema
}
\description{
Get the WebApi configuration for the WebApi instance.
}
\details{
Get the WebApi configuration for the WebApi instance.
}
\examples{
\dontrun{
#

getConfiguration(baseUrl = "http://server.org:80/WebAPI", sourceKeys = c('source1', 'source2'))
}

}
