% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webApiDataSources.R
\name{getDataSource}
\alias{getDataSource}
\title{Get the all the results from all Achilles reports for all provided data Sources}
\usage{
getDataSource(baseUrl, sourceKeys)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceKeys}{A list of source keys for a CDM instance in WebAPI, as defined in the
Configuration page.}
}
\value{
A list of object representing pre computed Achilles results for the list of sourceKey's
}
\description{
Get the all the results from all Achilles reports for all provided data Sources
}
\details{
Obtains the JSON expression from WebAPI for a list of dataSource, derived from given list of CDMschemas and returns list of tibble dataframes
}
\examples{
\dontrun{
# This will obtain pre-computed summary regarding dataSource

getDataSource(baseUrl = "http://server.org:80/WebAPI", cdmDatabaseSchemas = c('schema1','schema2')
}

}
