% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webApiDataSources.R
\name{getDataSourceReportForSourceKey}
\alias{getDataSourceReportForSourceKey}
\title{Get the results from individual Achilles reports by data Source definition expression}
\usage{
getDataSourceReportForSourceKey(baseUrl, sourceKey, reportType)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}

\item{reportType}{The precomputed report type in Atlas (Achilles report e.g. dashboard, datadensity, person etc)}
}
\value{
A list of object with pre computed Achilles results for the sourceKey and Achilles report
}
\description{
Get the results from individual Achilles reports by data Source definition expression
}
\details{
Obtain the JSON expression from WebAPI for a given dataSource
}
\examples{
\dontrun{
# This will obtain pre-computed summary regarding dataSource

getDataSourceReportForSourceKey(baseUrl = "http://server.org:80/WebAPI", dataSource = 'dataSource', reportType = 'dashboard')
}

}
