% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSources.R
\name{getDataSourceReportsForSourceKeys}
\alias{getDataSourceReportsForSourceKeys}
\title{Get the all the results from all Achilles reports for all provided data Sources}
\usage{
getDataSourceReportsForSourceKeys(baseUrl, listSourceKeys)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceKeys}{A list of data source key's for a CDM instance in WebAPI, as defined in the
Configuration page}
}
\value{
A list of object representing pre computed Achilles results for the list of sourceKey's
}
\description{
Get the all the results from all Achilles reports for all provided data Sources
}
\details{
Obtain the JSON expression from WebAPI for a given dataSource
}
\examples{
\dontrun{
# This will obtain pre-computed summary regarding dataSource

getDataSourceReportsForSourceKeys(baseUrl = "http://server.org:80/WebAPI", dataSources = c('dataSource1','dataSource2')
}

}
