% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDefinition.R
\name{getDefinition}
\alias{getDefinition}
\title{Get the definition for an id of chosen category in WebApi. \lifecycle{stable}}
\usage{
getDefinition(id, category, baseUrl)
}
\arguments{
\item{id}{An integer id representing the id that uniquely identifies a 
definition for the category in a WebApi instance.}

\item{category}{These are the categories in WebApi. The valid string options are 'conceptSet', 
'cohort', 'characterization', 'pathway, 'incidenceRate', 'estimation', 'prediction'.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
An R object representing the definition
}
\description{
Get the definition for an id of chosen category in WebApi. \lifecycle{stable}
}
\details{
Get the definition for an id of chosen category in WebApi. The return object will be a R
representation of the definition, that may be reconverted to JSON.
}
\examples{
\dontrun{
getDefinition(id = 13242, category = "cohort", baseUrl = "http://server.org:80/WebAPI")
}
}
