% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDefinitionsMetadata.R
\name{getDefinitionsMetadata}
\alias{getDefinitionsMetadata}
\title{Retrieve the meta data for WebApi definitions of a certain category \lifecycle{stable}}
\usage{
getDefinitionsMetadata(baseUrl, category)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{category}{These are the categories in WebApi. The valid string options are 'conceptSet', 
'cohort', 'characterization', 'pathway, 'incidenceRate', 'estimation', 'prediction'.}
}
\value{
A tibble of specification metadata.
}
\description{
Retrieve the meta data for WebApi definitions of a certain category \lifecycle{stable}
}
\details{
Obtains the meta data of WebApi specifications such as id, name, created/modified details, hash
object, etc for a certain category. The following function categories are supported. Concept-set,
Cohort-definition, Cohort-characterization, Pathway-analysis, Incidence rate (ir), estimation and
prediction. This function is useful to retrieve the current specifications.
}
\examples{
\dontrun{
getDefinitionsMetadata(baseUrl = "http://server.org:80/WebAPI")
}

}
