% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{getEstimation}
\alias{getEstimation}
\title{Get the estimation specification}
\usage{
getEstimation(baseUrl, estimationId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{estimationId}{The WebApi id for the estimation specification}
}
\value{
A list of R-objects with specifications for estimation
}
\description{
Get the estimation specification
}
\details{
Get an R list object with expression from WebAPI for a given estimation specification
}
\examples{
\dontrun{
getEstimation(baseUrl = "http://server.org:80/WebAPI", estimationId = 3434)
}

}
