% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R, R/Estimation.R
\name{getEstimationDefinition}
\alias{getEstimationDefinition}
\title{Get Estimation id definition.}
\usage{
getEstimationDefinition(baseUrl, estimationId)

getEstimationDefinition(baseUrl, estimationId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{estimationId}{The WebApi id for the estimation specification}

\item{EstimationId}{An integer id representing the id that uniquely identifies a Estimation
definition in a WebApi instance.}
}
\value{
An R object representing the Estimation definition

An R object.
}
\description{
Get Estimation id definition.

Get the estimation definition
}
\details{
Obtain the Estimation definition from WebAPI for a given Estimation id

Get an R object representing the estimation definition.
}
\examples{
\dontrun{
getEstimationDefinition(EstimationId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
\dontrun{
getEstimationDefinition(baseUrl = "http://server.org:80/WebAPI", estimationId = 3434)
}

}
