% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getEstimationDefinitionsMetaData}
\alias{getEstimationDefinitionsMetaData}
\title{Retrieve the meta data for Estimation definitions.}
\usage{
getEstimationDefinitionsMetaData(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A tibble of specification metadata for Estimation. Note: modifiedDate and createdDate are returned
as text/character.
}
\description{
Retrieve the meta data for Estimation definitions.
}
\details{
Obtains the meta data of WebApi specifications such as id, name, created/modified details, hash
object, etc. from WebApi for Estimation. This function is useful to retrieve the current Estimation
specifications.
}
\examples{
getEstimationDefinitionsMetaData(baseUrl = "http://server.org:80/WebAPI")

}
