% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{getEstimationSpecification}
\alias{getEstimationSpecification}
\title{Get the results from individual Achilles reports by data Source definition expression}
\usage{
getEstimationSpecification(baseUrl, estimationId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{estimationId}{The Atlas id for the estimation specification}
}
\value{
A list of objects with specifications for estimation
}
\description{
Get the results from individual Achilles reports by data Source definition expression
}
\details{
Obtain the JSON expression from WebAPI for a given dataSource
}
\examples{
\dontrun{
# This will obtain pre-computed summary regarding dataSource

getEstimationSpecification(baseUrl = "http://server.org:80/WebAPI", estimationId = 3434)
}

}
