% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceRateAnalysis.R
\name{getIncidenceRateAnalysis}
\alias{getIncidenceRateAnalysis}
\title{Get precomputed incidence rate analysis results from a specified Atlas incidence rate analyis ID
for the specified combination of Atlas targetId, Atlas outcomeId and Data Source.}
\usage{
getIncidenceRateAnalysis(baseUrl, irAnalysisId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{irAnalysisId}{The Atlas ID for incidence rate analysis.}
}
\value{
A list of Data Frame objects corresponding to the results of the Incidence Rate Analysis.
}
\description{
Get precomputed incidence rate analysis results from a specified Atlas incidence rate analyis ID
for the specified combination of Atlas targetId, Atlas outcomeId and Data Source.
}
\details{
Obtain a data frame with results
}
\examples{
\dontrun{
# This will obtain a list of Data Frame with results of the Incidence Rate Analysis:
getIncidenceRateAnalysis <- function(baseUrl "http://server.org:80/WebAPI",
                                         irAnalysisId = 296)
                                         )
}
                       
}
