% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceRateAnalysis.R
\name{getIncidenceRateAnalysisResults}
\alias{getIncidenceRateAnalysisResults}
\title{Get precomputed incidence rate analysis results from a specified Atlas incidence rate analyis ID
for the specified combination of Atlas targetId, Atlas outcomeId and Data Source.}
\usage{
getIncidenceRateAnalysisResults(baseUrl = NULL, irAnalysisId = NULL,
  sourceKey = NULL, targetId = NULL, outcomeId = NULL)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{irAnalysisId}{The Atlas ID for incidence rate analysis.}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the Atlas Configuration page.}

\item{targetId}{The Atlas ID for cohort definition corresponding to the target cohort.}

\item{outcomeId}{The Atlas ID for cohort definition corresponding to the outcome cohort.}
}
\value{
A list of Data Frame objects corresponding to the results of the Incidence Rate Analysis.
}
\description{
Get precomputed incidence rate analysis results from a specified Atlas incidence rate analyis ID
for the specified combination of Atlas targetId, Atlas outcomeId and Data Source.
}
\details{
Obtain a data frame with results
}
\examples{
\dontrun{
# This will obtain a list of Data Frame with results of the Incidence Rate Analysis:
getIncidenceRateAnalysisResults <- function(baseUrl "http://server.org:80/WebAPI",
                                         irAnalysisId = 296,
                                         sourceKey = 'YOUR_DATA_SOURCE',
                                         targetId =10817,
                                         outcomeId = 10873)
                                         )
}
                       
}
