% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R, R/IncidenceRate.R
\name{getIncidenceRateDefinition}
\alias{getIncidenceRateDefinition}
\title{Get IncidenceRate id definition.}
\usage{
getIncidenceRateDefinition(baseUrl, incidenceRateId)

getIncidenceRateDefinition(baseUrl, incidenceRateId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{incidenceRateId}{The WebApi ID for incidence rate analysis.}

\item{IncidenceRateId}{An integer id representing the id that uniquely identifies a IncidenceRate
definition in a WebApi instance.}
}
\value{
An R object representing the IncidenceRate definition

Get definitions for incident rate analysis
}
\description{
Get IncidenceRate id definition.

Get definitions for incident rate analysis
}
\details{
Obtain the IncidenceRate definition from WebAPI for a given IncidenceRate id

Returns an R-objects with definitions for incident rate analysis
}
\examples{
\dontrun{
getIncidenceRateDefinition(IncidenceRateId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
\dontrun{
 getIncidenceRateDefinition <- function(baseUrl "http://server.org:80/WebAPI",
                                         incidenceRateId = 296)
                                         )
}
}
