% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getIncidenceRateGenerationInformation}
\alias{getIncidenceRateGenerationInformation}
\title{Get generation information for IncidenceRate id.}
\usage{
getIncidenceRateGenerationInformation(incidenceRateId, sourceKey, baseUrl)
}
\arguments{
\item{incidenceRateId}{An integer id representing the id that uniquely identifies a IncidenceRate
definition in a WebApi instance.}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
An R object representing the IncidenceRate definition
}
\description{
Get generation information for IncidenceRate id.
}
\details{
Get generation (execution) information about IncidenceRate for a given combination of sourceKey and
incidenceRateId.
}
\examples{
\dontrun{
getIncidenceRateGenerationInformation(incidenceRateId = 13242,
                                      baseUrl = "http://server.org:80/WebAPI",
                                      sourceKey = "HCUP")
}
}
