% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceRate.R
\name{getIncidenceRateOutput}
\alias{getIncidenceRateOutput}
\title{Get output from incidence rate analysis.}
\usage{
getIncidenceRateOutput(
  baseUrl,
  incidenceRateId,
  sourceKey,
  targetCohortId,
  outcomeCohortId
)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{incidenceRateId}{The WebApi ID for incidence rate analysis.}

\item{sourceKey}{The sourceKey for the incident rate analysis.}

\item{targetCohortId}{The cohort id of the target cohort in the incident rate analysis.}

\item{outcomeCohortId}{The cohort id of the outcome cohort in the incidence rate analysis.}
}
\value{
A list with incident rate results summary, stratified results and treemap data.
}
\description{
Get output from incidence rate analysis.
}
\details{
Given a sourceKey, targetCohortId, outcomeCohortId combinations, this function will return the 
output of incidence rate analysis.
}
\examples{
\dontrun{
getIncidenceRateOutput <- function(baseUrl "http://server.org:80/WebAPI",
                                    incidenceRateId = 296,
                                    targetCohortId = 432423, 
                                    outcomeCohortId = 324,
                                    sourceKey = "HCUP"
                                   )
}
}
