% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceRate.R
\name{getIncidenceRateResults}
\alias{getIncidenceRateResults}
\title{Get results from incidence rate analysis.}
\usage{
getIncidenceRateResults(
  baseUrl,
  incidenceRateId,
  sourceKey,
  targetCohortId,
  outcomeCohortId
)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{incidenceRateId}{The Atlas ID for incidence rate analysis.}

\item{sourceKey}{The sourceKey for the incident rate analysis.}

\item{targetCohortId}{The cohort id of the target cohort in the incident rate analysis.}

\item{outcomeCohortId}{The cohort id of the outcome cohort in the incidence rate analysis.}
}
\value{
A list with incident rate results summary, stratified results and treemap data.
}
\description{
Get results from incidence rate analysis.
}
\details{
Given a sourceKey, targetCohortId, outcomeCohortId combinations, this function will return the 
results of incidence rate analysis.
}
\examples{
\dontrun{
getIncidenceRateResults <- function(baseUrl "http://server.org:80/WebAPI",
                                    incidenceRateId = 296,
                                    targetCohortId = 432423, 
                                    outcomeCohortId = 324,
                                    sourceKey = "HCUP"
                                   )
}
}
