% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceRate.R
\name{getIncidenceRateSpecification}
\alias{getIncidenceRateSpecification}
\title{Get definitions for incident rate analysis}
\usage{
getIncidenceRateSpecification(baseUrl, incidenceRateId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{incidenceRateId}{The WebApi ID for incidence rate analysis.}
}
\value{
Get definitions for incident rate analysis
}
\description{
Get definitions for incident rate analysis
}
\details{
Returns an R-objects with definitions for incident rate analysis
}
\examples{
\dontrun{
 getIncidenceRateSpecification <- function(baseUrl "http://server.org:80/WebAPI",
                                         incidenceRateId = 296)
                                         )
}
}
