% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getIncidenceRateSql}
\alias{getIncidenceRateSql}
\title{Get SQL query for IncidenceRate definition.}
\usage{
getIncidenceRateSql(incidenceRateDefinition, baseUrl)
}
\arguments{
\item{incidenceRateDefinition}{An R list object (not JSON) representing the IncidenceRate
definition. It is the output R expression object of list object
from \code{IncidenceRateDefinition}}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
An R object containing the SQL for IncidenceRate definition.
}
\description{
Get SQL query for IncidenceRate definition.
}
\details{
Given a valid IncidenceRate definition R-object (not JSON) this function will return the
parametrized SQL in OHDSI SQL dialect. This SQL then may be used along with OHDSI R-package
'SQLRender' to render/translate to target SQL dialect and parameters rendered.
}
\examples{
\dontrun{
getIncidenceRateSql(IncidenceRateDefinition = (getIncidenceRateDefinition(cohortId = 13242,
                                                                          baseUrl = baseUrl)),
                    baseUrl = "http://server.org:80/WebAPI")
}
}
