% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vocabulary.R
\name{getMappedStandardConcepts}
\alias{getMappedStandardConcepts}
\title{Get Mapped Standard concepts for a source concept}
\usage{
getMappedStandardConcepts(baseUrl, sourceConceptId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceConceptId}{A source concept id for which we are to find the mapped standard concepts}
}
\value{
A list of standard concepts mapped to the source concept id
}
\description{
Get Mapped Standard concepts for a source concept
}
