% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vocabulary.R
\name{getMedDraFromSourceCode}
\alias{getMedDraFromSourceCode}
\title{Get MedDRA terms from 1 source code}
\usage{
getMedDraFromSourceCode(baseUrl, sourceCode, sourceVocabularyId)
}
\arguments{
\item{baseUrl}{The WebAPI endpoint}

\item{sourceCode}{A single source code}

\item{sourceVocabularyId}{The name of the source vocabulary id for the source code (e.g. ICD9CM, ICD10CM)}
}
\description{
Get MedDRA terms from 1 source code
}
\details{
Obtain MedDRA Preferred Terms that are direct ancestors for a data frame of
condition source codes. Note: you must have MedDRA terms loaded in your OMOP Vocabulary.
}
