% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vocabulary.R
\name{getMedDraFromSourceCodeDf}
\alias{getMedDraFromSourceCodeDf}
\title{Get MedDRA terms from source code data frame}
\usage{
getMedDraFromSourceCodeDf(baseUrl, sourceCodeDf, maxThreads = 10)
}
\arguments{
\item{baseUrl}{The WebAPI endpoint}

\item{sourceCodeDf}{A data frame consisting of: SOURCE_CODE and SOURCE_VOCABULARY_ID}

\item{maxThreads}{(OPTIONAL) The maximum number of parallel requests to make to WebAPI. 
Default is 10. It is advisable to not use too many.}
}
\description{
Get MedDRA terms from source code data frame
}
\details{
Obtain MedDRA Preferred Terms that are direct ancestors for a data frame of
condition source codes. Note: you must have MedDRA terms loaded in your OMOP Vocabulary.
}
