% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{getMyPermissions}
\alias{getMyPermissions}
\title{Get my permissions}
\usage{
getMyPermissions(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
A tibble with user ID, user name, and permission information when available.
}
\description{
This function returns the authorized users permission for
definitions like cohort, incidence rate, characterization..
}
\examples{
\dontrun{
getUserInformations(baseUrl = "http://server.org:80/WebAPI")
}
}
