% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getPathwayDefinition}
\alias{getPathwayDefinition}
\title{Get Pathway id definition.}
\usage{
getPathwayDefinition(pathwayId, baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{PathwayId}{An integer id representing the id that uniquely identifies a Pathway definition
in a WebApi instance.}
}
\value{
An R object representing the Pathway definition
}
\description{
Get Pathway id definition.
}
\details{
Obtain the Pathway definition from WebAPI for a given Pathway id
}
\examples{
\dontrun{
getPathwayDefinition(PathwayId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
