% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getPathwayDefinitionsMetaData}
\alias{getPathwayDefinitionsMetaData}
\title{Retrieve the meta data for Pathway definitions.}
\usage{
getPathwayDefinitionsMetaData(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A tibble of specification metadata for Pathway. Note: modifiedDate and createdDate are returned as
text/character.
}
\description{
Retrieve the meta data for Pathway definitions.
}
\details{
Obtains the meta data of WebApi specifications such as id, name, created/modified details, hash
object, etc. from WebApi for Pathway. This function is useful to retrieve the current Pathway
specifications.
}
\examples{
\dontrun{
getPathwayDefinitionsMetaData(baseUrl = "http://server.org:80/WebAPI")
}
}
