% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getPathwayGenerationInformation}
\alias{getPathwayGenerationInformation}
\title{Get generation information for Pathway id.}
\usage{
getPathwayGenerationInformation(pathwayId, baseUrl)
}
\arguments{
\item{pathwayId}{An integer id representing the id that uniquely identifies a 
Pathway definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
An R object representing the Pathway definition
}
\description{
Get generation information for Pathway id.
}
\details{
Get generation (execution) information about Pathway for a pathwayId.
}
\examples{
\dontrun{
getPathwayGenerationInformation(pathwayId = 13242, 
                                             baseUrl = "http://server.org:80/WebAPI")
}
}
