% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getPathwayResults}
\alias{getPathwayResults}
\title{Get results for a Pathway Id.}
\usage{
getPathwayResults(pathwayId, baseUrl)
}
\arguments{
\item{pathwayId}{An integer id representing the id that uniquely identifies a pathway analysis definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
An R object with results.
}
\description{
Get results for a Pathway Id.
}
\details{
Get the results for Pathway id.
}
\examples{
\dontrun{
getPathwayResults(pathwayId = 342, baseUrl = "http://server.org:80/WebAPI")
}
}
