% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonProfile.R
\name{getPersonProfile}
\alias{getPersonProfile}
\title{Get person profile data}
\usage{
getPersonProfile(baseUrl, sourceKey, personId, cohortId = NULL)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}

\item{personId}{The personId of the person whoose profile is being reviewed}

\item{cohortId}{<OPTIONAL> The cohortId of the cohort that the person belongs to}
}
\value{
A tibble data frame object
}
\description{
Get person profile data
}
\details{
Get a R object with person profile data. This function may be used for 
visualizing a patients profile in tables or visualization.
}
\examples{
\dontrun{#'
getPersonProfile(baseUrl = "http://server.org:80/WebAPI", 
                 sourceKey = 'sourceKey',
                 personId = 342342)
}

}
