% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R, R/Prediction.R
\name{getPredictionDefinition}
\alias{getPredictionDefinition}
\title{Get Prediction id definition.}
\usage{
getPredictionDefinition(predictionId, baseUrl)

getPredictionDefinition(predictionId, baseUrl)
}
\arguments{
\item{predictionId}{An integer id representing the id that uniquely identifies a prediction definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{PredictionId}{An integer id representing the id that uniquely identifies a Prediction
definition in a WebApi instance.}
}
\value{
An R object representing the Prediction definition

An R object representing the cohort definition
}
\description{
Get Prediction id definition.

Get a prediction definition
}
\details{
Obtain the Prediction definition from WebAPI for a given Prediction id

Obtain the prediction definition from WebAPI for a given prediction id
}
\examples{
\dontrun{
getPredictionDefinition(PredictionId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
\dontrun{
getPredictionDefinition(predictionId = 282, baseUrl = "http://server.org:80/WebAPI")
}

}
