% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getPredictionDefinitionsMetaData}
\alias{getPredictionDefinitionsMetaData}
\title{Get the meta data for Prediction definitions. \lifecycle{stable}}
\usage{
getPredictionDefinitionsMetaData(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
A tibble of specification metadata for Prediction. Note: modifiedDate and createdDate are returned
as text/character.
}
\description{
Get the meta data for Prediction definitions. \lifecycle{stable}
}
\details{
Get the meta data of WebApi specifications such as id, name, created/modified details, hash object,
etc. from WebApi for Prediction. This function is useful to retrieve the current Prediction
specifications.
}
\examples{
\dontrun{
getPredictionDefinitionsMetaData(baseUrl = "http://server.org:80/WebAPI")
}
}
