% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{getPriorityVocabKey}
\alias{getPriorityVocabKey}
\title{Get Priority Vocab Source Key}
\usage{
getPriorityVocabKey(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".}
}
\value{
A string with the source key of the default OMOP Vocab in Atlas.
}
\description{
Get Priority Vocab Source Key
}
\details{
Obtains the source key of the default OMOP Vocab in Atlas.
}
