% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{getPriorityVocabularyKey}
\alias{getPriorityVocabularyKey}
\title{Get Priority Vocabulary Source Key}
\usage{
getPriorityVocabularyKey(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A string.
}
\description{
Get Priority Vocabulary Source Key
}
\details{
Obtains the source key of the default OMOP Vocabulary in WebApi.
}
