% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vocabulary.R
\name{getRelatedConcepts}
\alias{getRelatedConcepts}
\title{Get Related Concepts for a concept id}
\usage{
getRelatedConcepts(baseUrl, conceptId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{conceptId}{A concept id for which we are to find the related concepts}
}
\description{
Get Related Concepts for a concept id
}
