% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getSetExpressionConceptIds}
\alias{getSetExpressionConceptIds}
\title{Get Concepts from a Concept Set Expression}
\usage{
getSetExpressionConceptIds(baseUrl, expression, vocabSourceKey = NULL)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{expression}{A JSON string that represents the concept set expression}

\item{vocabSourceKey}{The source key of the Vocabulary. By default, the priority Vocabulary is
used.}
}
\value{
A list of concept ids
}
\description{
Get Concepts from a Concept Set Expression
}
\details{

}
\examples{
\dontrun{
# This will obtain the concept ids from a concept set expression:

getSetExpressionConceptIds(baseUrl = "http://server.org:80/WebAPI",
                           expression = someJsonExpression)
}

}
