% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vocabulary.R
\name{getSourceConcept}
\alias{getSourceConcept}
\title{Get Source Concept Id from Source Code}
\usage{
getSourceConcept(baseUrl, sourceCode, sourceVocabularyId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceCode}{A source code}

\item{sourceVocabularyId}{The vocabulary id that the source code belongs to}
}
\value{
A list object defining the source concept
}
\description{
Get Source Concept Id from Source Code
}
