% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concept.R
\name{getSourceConcepts}
\alias{getSourceConcepts}
\title{Get source concepts that map to standard concepts}
\usage{
getSourceConcepts(
  conceptIds,
  baseUrl,
  vocabSourceKey = NULL,
  snakeCaseToCamelCase = TRUE
)
}
\arguments{
\item{conceptIds}{A list of concept IDs referring to standard concepts.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{vocabSourceKey}{The source key of the Vocabulary. By default, the priority Vocabulary is
used.}

\item{snakeCaseToCamelCase}{Should the column names of the result be converted to camelCase?}
}
\value{
A tibble containing information on the source concepts.
}
\description{
Get source concepts that map to standard concepts
}
\examples{
\dontrun{
conceptSet <- getConceptSet(conceptSetId = 282, baseUrl = "http://server.org:80/WebAPI")
conceptIds <- resolveConceptSet(conceptSet = conceptSet,
                                baseUrl = "http://server.org:80/WebAPI")
sourceConcepts <- getSourceConcepts(conceptIds = conceptIds, 
                                    baseUrl = "http://server.org:80/WebAPI")
}

}
