% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{getSourceKeyConfiguration}
\alias{getSourceKeyConfiguration}
\title{Get configuration information for source keys.}
\usage{
getSourceKeyConfiguration(baseUrl, sourceKeys)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceKeys}{A list of string names for the source keys as in
configuration page of Atlas.}
}
\value{
A list of two objects called native and parsed. Native is the json 
specification, and parsed is the simplified version returned as a data.frame
object. The data.frame object may be used for creating reports.
}
\description{
Get configuration information for source keys.
}
\details{
Source keys are commonly used as aliases to identify a data source. 
A data source is uniquely identified a combination of sourceName and/or CDM name.
sourceName/CDM name may also have version information. 
WebApi calls rely on sourceKey for query. But relying on aliases
is not reliable, as the underlying data source may change.
This function, will give the configuration information for data sources
given source keys. It may be used to unambigously identify the data source
in a study.
}
\examples{
\dontrun{
#

getSourceKeyConfiguration(baseUrl = "http://server.org:80/WebAPI", 
                                     sourceKeys = c('source1', 'source2')
                                   )$parsed
would return a data.frame with items like sourceId, sourceName, sourceDialect,
sourceKey, priority, CDM, CEM.
}

}
