% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{getUserIdFromUserName}
\alias{getUserIdFromUserName}
\title{Get the user id for a given user name. \lifecycle{stable}}
\usage{
getUserIdFromUserName(baseUrl, userName)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{userName}{A user name that is used to authenticate with WebApi.}
}
\description{
Get the user id for a given user name. \lifecycle{stable}
}
\details{
The function will call the WebApi. to assign permissions to a cohort definition for the given userId.
}
\examples{
\dontrun{
getUserIdFromUserName(baseUrl, userName)
}
}
