% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{getUserInformation}
\alias{getUserInformation}
\title{Get user information}
\usage{
getUserInformation(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
A tibble with user ID, user name, and permission information when available.
}
\description{
This function returns all public users in the security enabled webapi, and their
public permissions. This function only works in security enabled webapi.
}
\examples{
\dontrun{
getUserInformations(baseUrl = "http://server.org:80/WebAPI")
}
}
