% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permissions.R
\name{getUserPermission}
\alias{getUserPermission}
\title{Get user permission}
\usage{
getUserPermission(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
A dataframe with user ID, user name, and permission information
}
\description{
This function only applies to WebAPI instances with security enabled. If security is not
emabled then it will return an error.
}
\examples{
\dontrun{
getUserPermissions(baseUrl = "http://server.org:80/WebAPI")
}
}
