% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{getUsersFromRole}
\alias{getUsersFromRole}
\title{Obtains users for a specific WebAPI user role. Requires security enabled WebAPI
\lifecycle{experimental}}
\usage{
getUsersFromRole(baseUrl, roleId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{roleId}{The role id as defined in WebAPI sec_role table}
}
\value{
A data frame of user information for the WebAPI role.
}
\description{
Obtains users for a specific WebAPI user role. Requires security enabled WebAPI
\lifecycle{experimental}
}
\details{
For a security enabled WebAPI, obtains the user names for a specified Atlas user role. Role Id is
set in WebAPI sec_role table
}
\examples{
\dontrun{
getUsersFromRole(baseUrl = "http://server.org:80/WebAPI", roleId = 10)
}
}
