% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{getWebApiDefinitionsMetadata}
\alias{getWebApiDefinitionsMetadata}
\title{Retrieve the meta data of all WebApi definitions}
\usage{
getWebApiDefinitionsMetadata(baseUrl, categories)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{categories}{These are the categories in WebApi. The valid options are conceptSet, 
cohort, incidenceRate, estimation, prediction, characterization, pathway.}
}
\value{
A tibble of specification metadata. Note: modifiedDate and createdDate are
returned as text/character.
}
\description{
Retrieve the meta data of all WebApi definitions
}
\details{
Obtains the meta data of WebApi specifications such as id, name, created/modified 
details, hash object, etc. The following function categories are supported. 
Concept-set, Cohort-definition, Cohort-characterization, Pathway-analysis, Incidence rate (ir), 
estimation and prediction. This function is useful to retrieve the current specifications.
}
\examples{
\dontrun{
getWebApiDefinitionsMetadata(baseUrl = "http://server.org:80/WebAPI")
}

}
