% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{insertCohortDefinitionSetInPackage}
\alias{insertCohortDefinitionSetInPackage}
\title{Insert a set of cohort definitions into package}
\usage{
insertCohortDefinitionSetInPackage(
  fileName = "inst/settings/CohortsToCreate.csv",
  baseUrl,
  jsonFolder = "inst/cohorts",
  sqlFolder = "inst/sql/sql_server",
  rFileName = "R/CreateCohorts.R",
  insertTableSql = TRUE,
  insertCohortCreationR = TRUE,
  generateStats = FALSE,
  packageName
)
}
\arguments{
\item{fileName}{Name of a CSV file specifying the cohorts to insert. See details for
the expected file format.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{jsonFolder}{Path to the folder where the JSON representations will be saved.}

\item{sqlFolder}{Path to the folder where the SQL representations will be saved.}

\item{rFileName}{Name of R file to generate when \code{insertCohortCreationR = TRUE}.}

\item{insertTableSql}{Should the SQL for creating the cohort table be inserted into the
package as well? This file will be called CreateCohortTable.sql.}

\item{insertCohortCreationR}{Insert R code that will create the cohort table and instantiate the
cohorts? This will create a file called R/CreateCohorts.R containing
a function called \code{.createCohorts}.}

\item{generateStats}{Should cohort inclusion rule statistics be created?}

\item{packageName}{The name of the package (only needed when inserting the R code as
well).}
}
\description{
Insert a set of cohort definitions into package
}
\details{
The CSV file should have at least the following fields: \describe{ \item{atlasId}{The cohort ID in
ATLAS.} \item{cohortId}{The cohort ID that will be used when instantiating the cohort (can be
different from atlasId).} \item{name}{The name to be used for the cohort. This name will be used to
generate file names, so please use letters and numbers only (no spaces).} }
}
