% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{invokeCharacterizationGeneration}
\alias{invokeCharacterizationGeneration}
\title{Invoke generation of Characterization id.}
\usage{
invokeCharacterizationGeneration(characterizationId, baseUrl, sourceKey)
}
\arguments{
\item{characterizationId}{An integer id representing the id that uniquely identifies a
Characterization definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}
}
\value{
A tibble with job status information.
}
\description{
Invoke generation of Characterization id.
}
\details{
Invoke the generation of Characterization id in the WebApi.
}
\examples{
\dontrun{
invokeCharacterizationGeneration(characterizationId = 13242,
                                 baseUrl = "http://server.org:80/WebAPI",
                                 sourceKey = "HCUP")
}
}
