% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{invokeCohortDefinition}
\alias{invokeCohortDefinition}
\title{Invoke generation of Cohort id.}
\usage{
invokeCohortDefinition(cohortId, baseUrl, sourceKey)
}
\arguments{
\item{cohortId}{An integer id representing the id that uniquely identifies a Cohort definition in
a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}
}
\value{
An R object representing the Cohort definition
}
\description{
Invoke generation of Cohort id.
}
\details{
Invoke the generation of Cohort id in the WebApi.
}
\examples{
\dontrun{
invokeCohortDefinition(CohortId = 13242,
                       baseUrl = "http://server.org:80/WebAPI",
                       sourceKey = "HCUP")
}
}
