% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{invokeCohortGeneration}
\alias{invokeCohortGeneration}
\title{Invoke the generation of cohort}
\usage{
invokeCohortGeneration(baseUrl, sourceKey, cohortId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}

\item{cohortId}{An integer id representing the id that uniquely identifies a cohort definition
in a WebApi instance.}
}
\description{
Invoke the generation of cohort
}
\details{
Invokes the generation of a cohort for a given combination of sourceKey and
cohortId. Use \code{getCohortGenerationStatus} to check the progress.
}
