% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{invokeCohortSetGeneration}
\alias{invokeCohortSetGeneration}
\title{Invoke the generation of a set of cohort definitions}
\usage{
invokeCohortSetGeneration(baseUrl, sourceKeys, definitionIds)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceKeys}{A list of CDM source keys. These can be found in Atlas -> Configure.}

\item{definitionIds}{A list of cohort definition Ids}
}
\description{
Invoke the generation of a set of cohort definitions
}
\details{
Invokes the generation of a set of cohort definitions across a set of CDMs set up in WebAPI. Use
\code{getCohortGenerationStatuses} to check the progress of the set.
}
