% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvokeGeneration.R
\name{invokeGeneration}
\alias{invokeGeneration}
\title{Invoke generation. \lifecycle{stable}}
\usage{
invokeGeneration(id, baseUrl, sourceKey, category)
}
\arguments{
\item{id}{An integer id representing the id that uniquely identifies a 
definition for the category in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}

\item{category}{These are the categories in WebApi. The valid string options are 'conceptSet', 
'cohort', 'characterization', 'pathway, 'incidenceRate', 'estimation', 'prediction'.}
}
\value{
Error message if invoke failed.
}
\description{
Invoke generation. \lifecycle{stable}
}
\details{
Invoke generation (execution) information.
}
\examples{
\dontrun{
invokeGeneration(id = 13242, category = "cohort", baseUrl = "http://server.org:80/WebAPI")
}
}
