% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{invokeIncidenceRateGeneration}
\alias{invokeIncidenceRateGeneration}
\title{Invoke generation of IncidenceRate id.}
\usage{
invokeIncidenceRateGeneration(incidenceRateId, baseUrl, sourceKey)
}
\arguments{
\item{incidenceRateId}{An integer id representing the id that uniquely identifies a IncidenceRate
definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}
}
\value{
A tibble with job status information.
}
\description{
Invoke generation of IncidenceRate id.
}
\details{
Invoke the generation of IncidenceRate id in the WebApi.
}
\examples{
\dontrun{
invokeIncidenceRateGeneration(incidenceRateId = 13242,
                              baseUrl = "http://server.org:80/WebAPI",
                              sourceKey = "HCUP")
}
}
