% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{invokePathwayDefinition}
\alias{invokePathwayDefinition}
\title{Invoke generation of Pathway id.}
\usage{
invokePathwayDefinition(pathwayId, baseUrl, sourceKey)
}
\arguments{
\item{pathwayId}{An integer id representing the id that uniquely identifies a Pathway definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}
}
\value{
An R object representing the Pathway definition
}
\description{
Invoke generation of Pathway id.
}
\details{
Invoke the generation of Pathway id in the WebApi.
}
\examples{
\dontrun{
invokePathwayDefinition(PathwayId = 13242,
                        baseUrl = "http://server.org:80/WebAPI",
                        sourceKey = "HCUP")
}
}
