% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{isValidCharacterizationId}
\alias{isValidCharacterizationId}
\title{Check if Characterization id is valid.}
\usage{
isValidCharacterizationId(characterizationIds, baseUrl)
}
\arguments{
\item{characterizationIds}{A list of integer id(s) of the Characterization to be tested for
validity.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A logical vector indicating which of the ids are valid.
}
\description{
Check if Characterization id is valid.
}
\details{
Checks if a set of id for a Characterization is valid, i.e. checks if all the ids exists in the
WebApi i.e. valid.
}
\examples{
\dontrun{
isValidCharacterizationId(characterizationIds = c(13242, 3423, 34),
                          baseUrl = "http://server.org:80/WebAPI")
}
}
