% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{isValidCohortId}
\alias{isValidCohortId}
\title{Check if Cohort id is valid.}
\usage{
isValidCohortId(cohortIds, baseUrl)
}
\arguments{
\item{cohortIds}{A list of integer id(s) of the Cohort to be tested for validity.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A logical vector indicating which of the ids are valid.
}
\description{
Check if Cohort id is valid.
}
\details{
Checks if a set of id for a Cohort is valid, i.e. checks if all the ids exists in the WebApi i.e.
valid.
}
\examples{
\dontrun{
isValidCohortId(cohortIds = c(13242, 3423, 34), baseUrl = "http://server.org:80/WebAPI")
}
}
