% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{isValidSourceKey}
\alias{isValidSourceKey}
\title{Check if source key is valid.}
\usage{
isValidSourceKey(sourceKeys, baseUrl)
}
\arguments{
\item{sourceKeys}{The source key(s) for a CDM instance in WebAPI, as defined in the Configuration
page.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
A logical vector indicating if an ID is valid.
}
\description{
Check if source key is valid.
}
\details{
Checks if a set of sourceKey(s) are valid, i.e. checks if all the sourceKey(s) exists in the WebApi
i.e. valid.
}
\examples{
\dontrun{
isValidSourceKey(sourceKeys = c("HCUP", "CCA"),
                 baseUrl = "http://server.org:80/WebAPI",
                 category = "cohort")
}
}
