% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{postAtlasDefinition}
\alias{postAtlasDefinition}
\title{Post a atlas study definition}
\usage{
postAtlasDefinition(baseUrl, name, definitionType = "cohort", jsonExpression)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{name}{A valid name for atlas study definition.Please make sure that the
name is unique in the target webapi - otherwise, an error might be thrown.}

\item{definitionType}{Currently only 'cohort' is supported and this corresponds to 'Cohort-definition'.}

\item{jsonExpression}{A valid JSON expression for an atlas study definition. Note: this is
expected to be a JSON object. Atlas names should be specified seperately
and are not part of this JSON expression. AtlasId's will be assigned by WebApi.}
}
\value{
This function will return a dataframe object with one row
                        describing the posted atlas study definitions and its details.
                        See \code{\link{getAtlasDefinitionsDetails}}. 
                        If unsuccessful a STOP message will be shown.
}
\description{
Post a atlas study definition
}
\details{
Posts a definition JSON into WebAPI
}
\examples{
\dontrun{
validJsonExpression <- getCohortDefinition(baseUrl = baseUrl, cohortId = 15873)
validJsonExpression <- RJSONIO::toJSON(cohortDefinition$expression)
postAtlasDefinition(name = 'new atlas cohort name', 
                     baseUrl = "http://server.org:80/WebAPI", 
                     jsonExpression = validJsonExpression)
}
}
