% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{postCohortDefinition}
\alias{postCohortDefinition}
\title{Post Cohort definition. \lifecycle{maturing}}
\usage{
postCohortDefinition(name, cohortDefinition, baseUrl)
}
\arguments{
\item{name}{A valid name for the definition. WebApi will use this name (if valid) as
the name of the definition. WebApi checks for validity, such as
uniqueness, absence of unacceptable character etc. An error might be
thrown.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{CohortDefinition}{An R list object containing the expression for the specification. This
will be converted to JSON expression by function and posted into the
WebApi. Note: only limited checks are performed in R to check the validity
of this expression.}
}
\value{
This function will return a dataframe object with one row describing the posted WebApi expression
and its details. If unsuccessful a STOP message will be shown.
}
\description{
Post Cohort definition. \lifecycle{maturing}
}
\details{
Post Cohort definition to WebAPI
}
\examples{
\dontrun{
postCohortDefinition(name = "new valid name",
                     cohortDefinition = definition,
                     baseUrl = "http://server.org:80/WebAPI")
}
}
