% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{postConceptSetDefinition}
\alias{postConceptSetDefinition}
\title{Post ConceptSet definition. \lifecycle{maturing}}
\usage{
postConceptSetDefinition(
  name,
  conceptSetDefinition,
  baseUrl,
  duplicateNames = "error"
)
}
\arguments{
\item{name}{A valid name for the definition. WebApi will use this name (if valid)
as the name of the definition. WebApi checks for validity, such as
uniqueness, absence of unacceptable character etc. An error might be
thrown.}

\item{conceptSetDefinition}{An R list object containing the expression for the specification. This
will be converted to JSON expression by function and posted into the
WebApi. Note: only limited checks are performed in R to check the
validity of this expression.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{duplicateNames}{How to handle importing a definition with a name that already exists
in ATLAS. 'error' will throw an error, 'overwrite' will attempt to
overwrite the existing definition, 'rename' will append the new
defintion name with (1) until the name is unique}
}
\value{
This function will return a dataframe object with one row describing the posted WebApi expression
and its details. If unsuccessful a STOP message will be shown.
}
\description{
Post ConceptSet definition. \lifecycle{maturing}
}
\details{
Post ConceptSet definition to WebAPI
}
\examples{
\dontrun{
postConceptSetDefinition(name = "new valid name",
                         conceptSetDefinition = definition,
                         baseUrl = "http://server.org:80/WebAPI")
}
}
