% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{postDefinition}
\alias{postDefinition}
\title{Post a definition into WebApi}
\usage{
postDefinition(baseUrl, name, category, object)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{name}{A valid name for the definition. WebApi will use this name (if valid) as
the name of the definition. WebApi checks for validity,
such as uniqueness, unaccepted character etc. An error might be thrown.}

\item{category}{The category of expression in WebApi. Currently only 'cohort' is supported,but
we will support posting the definitions of Concept-set, Cohort-definition, 
Cohort-characterization, Pathway-analysis, Incidence rate (ir), 
estimation and prediction. They should be referenced using string 
'conceptSet','cohort', 'characterization', 'pathway', 'incidenceRate', 
'estimation','prediction' only.}

\item{object}{An R list object containing the expression for the specification. 
This will be converted to JSON by function and posted into the WebApi.
Note: only limited checks are performed in R to check the validity of this
expression.}
}
\value{
This function will return a dataframe object with one row
                   describing the posted WebApi expression and its details.
                   If unsuccessful a STOP message will be shown.
}
\description{
Post a definition into WebApi
}
\details{
Post a definition into WebAPI
}
\examples{
\dontrun{
validJsonExpression <- getCohortDefinition(baseUrl = baseUrl, 
                                           cohortId = 15873)$expression
postDefinition(name = 'new name for expression in target',
                  baseUrl = "http://server.org:80/WebAPI",
                  object = validJsonExpression,
                  category = 'cohort')
}
}
