% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{postSpecification}
\alias{postSpecification}
\title{Post a specification expression into WebApi}
\usage{
postSpecification(baseUrl, name, type = "cohort", object)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{name}{A valid name for the expression. WebApi will use this name (if valid) as
the name of the specification. WebApi checks for validity,
such as uniqueness, unaccepted character etc. An error might be thrown.}

\item{type}{The type of expression in WebApi. Currently only 'cohort' is supported 
to refer cohort definition specification expression.}

\item{object}{An R list object containing the expression for the specification. 
This will be converted to JSON by function and posted into the WebApi.
Note: only limited checks are performed in R to check the validity of this
expression.}
}
\value{
This function will return a dataframe object with one row
                   describing the posted atlas study definitions and its details.
                   If unsuccessful a STOP message will be shown.
                   See \code{\link{getAtlasDefinitionsDetails}}.
}
\description{
Post a specification expression into WebApi
}
\details{
Post a specification expression into WebAPI
}
\examples{
\dontrun{
validJsonExpression <- getCohortDefinition(baseUrl = baseUrl, cohortId = 15873)$expression
postAtlasDefinition(name = 'new name for expression in target', 
                     baseUrl = "http://server.org:80/WebAPI", 
                     jsonExpression = validJsonExpression)
}
}
