% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{renameDfNamesFromSnakeCaseToCamelCase}
\alias{renameDfNamesFromSnakeCaseToCamelCase}
\title{Renames all columns in Data Frame from snake_Case to camelCase}
\usage{
renameDfNamesFromSnakeCaseToCamelCase(df)
}
\arguments{
\item{df}{The dataFrame that needs renaming}
}
\value{
A data frame with snake_Case names, changed to camelCase
}
\description{
Renames all columns in Data Frame from snake_Case to camelCase
}
\details{
Renames all columns in Data Frame from snake_Case to camelCase
}
