% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{resolveConceptSet}
\alias{resolveConceptSet}
\title{Resolve a concept set to the included standard concept IDs}
\usage{
resolveConceptSet(conceptSet, baseUrl, vocabSourceKey = NULL)
}
\arguments{
\item{conceptSet}{A concept set as obtained through the \code{\link{getConceptSet}} function.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{vocabSourceKey}{The source key of the Vocabulary. By default, the priority Vocabulary is
used.}
}
\value{
A list of standard concept ids
}
\description{
Resolve a concept set to the included standard concept IDs
}
\examples{
\dontrun{
conceptSet <- getConceptSet(conceptSetId = 282, baseUrl = "http://server.org:80/WebAPI")
conceptIds <- resolveConceptSet(conceptSet = conceptSet,
                                baseUrl = "http://server.org:80/WebAPI")
}

}
