% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{resolveConceptSetId}
\alias{resolveConceptSetId}
\title{Get resolved concept set expression}
\usage{
resolveConceptSetId(baseUrl, setId, formatName = FALSE)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{setId}{The concept set id in Atlas.}

\item{formatName}{Should the name be formatted to remove prefixes and underscores?}
}
\value{
A list of data frame objects with concept set expression, included conceptset, mapped conceptsets.
}
\description{
Get resolved concept set expression
}
\details{
Given a concept set JSON expression, return the included and/or mapped concept details
}
\examples{
\dontrun{
# This will obtain resolved concept set expression:

resolveConceptSet(baseUrl = "http://server.org:80/WebAPI", setId = 213423)
}

}
