% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROhdsiWebApi.R
\docType{package}
\name{ROhdsiWebApi-package}
\alias{ROhdsiWebApi}
\alias{ROhdsiWebApi-package}
\title{ROhdsiWebApi: Interacting With an OHDSI WebApi Instance}
\description{
ROhdsiWebApi is an R based interface to OHDSI WebAPI services, and performs GET/PULL/POST/DELETE calls via the WebApi. 
  All objects starting from R or output to R - are analysis ready R-objects like list and data.frame. 
  The package handles the intermediary steps by converting R-objects to JSON and vice versa.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/ROhdsiWebApi/}
  \item \url{https://github.com/OHDSI/ROhdsiWebApi}
  \item Report bugs at \url{https://github.com/OHDSI/ROhdsiWebApi/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Authors:
\itemize{
  \item Martijn Schuemie
  \item Jamie Gilbert \email{JGilber2@its.jnj.com}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
