% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{securityEnabled}
\alias{securityEnabled}
\title{Is WebAPI using authentication/authorization}
\usage{
securityEnabled(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
TRUE or FALSE
}
\description{
Is WebAPI using authentication/authorization
}
\examples{
\dontrun{
securityEnabled(baseUrl = "http://server.org:80/WebAPI")
}
}
