% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auth.R
\name{setAuthHeader}
\alias{setAuthHeader}
\title{Manually set the authorization http header for a WebAPI baseUrl}
\usage{
setAuthHeader(baseUrl, authHeader)
}
\description{
Manually set the authorization http header for a WebAPI baseUrl
}
