% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{setCohortPermission}
\alias{setCohortPermission}
\title{Assign cohort definition permissions to user. \lifecycle{experimental}}
\usage{
setCohortPermission(baseUrl, cohortId, userId, permission = "READ")
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{cohortId}{The ID of the cohort definition}

\item{userId}{The id to grant permission. You can find out the id from userName 
as getUserIdFromUserName(baseUrl = baseUrl, userName = "myId")}

\item{permission}{The permission type to assign.  Can be "READ" or "WRITE". Defaults to "WRITE".}
}
\description{
Assign cohort definition permissions to user. \lifecycle{experimental}
}
\details{
The function will call the webservice to assign permissions to a cohort definition for the given userId.
}
\examples{
\dontrun{
setCohortPermission(baseUrl, cohortId, userId = getUserIdFromUserName(baseUrl = baseUrl, userName = "myId"), permission = "READ")
}
}
