% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{updateCharacterizationDefinition}
\alias{updateCharacterizationDefinition}
\title{Update a Characterization definition. \lifecycle{maturing}}
\usage{
updateCharacterizationDefinition(
  characterizationDefinition,
  baseUrl,
  displayWarnings = TRUE
)
}
\arguments{
\item{characterizationDefinition}{An R list object containing the expression for the
specification. Must include id, name and expression. This will
be converted to JSON expression by function and posted into the
WebApi. The definition will be checked against the WebApi
instance for errors}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{displayWarnings}{Display warnings returned by WebApi check}
}
\description{
Update a Characterization definition. \lifecycle{maturing}
}
\details{
Update a Characterization definition.
}
\examples{
\dontrun{
definition <- getCharacterizationDefinition(id = 13242,
                                            baseUrl = "http://server.org:80/WebAPI",
                                            category = characterization)
definition$name <- "My new name for this"
updateCharacterization(characterizationDefinition, baseUrl, category = "cohort")
}
}
