% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{updateCohortDefinition}
\alias{updateCohortDefinition}
\title{Update a Cohort definition. \lifecycle{maturing}}
\usage{
updateCohortDefinition(cohortDefinition, baseUrl)
}
\arguments{
\item{cohortDefinition}{An R list object containing the expression for the specification. Must
include id, name and expression. This will be converted to JSON expression
by function and posted into the WebApi. The definition will be checked
against the WebApi instance for errors}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\description{
Update a Cohort definition. \lifecycle{maturing}
}
\details{
Update a Cohort definition.
}
\examples{
\dontrun{
definition <- getCohortDefinition(id = 13242,
                                  baseUrl = "http://server.org:80/WebAPI",
                                  category = cohort)
definition$name <- "My new name for this"
updateCohort(cohortDefinition, baseUrl, category = "cohort")
}
}
