% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPostDeleteDefinition.R
\name{updateDefinition}
\alias{updateDefinition}
\title{Update definition \lifecycle{maturing}}
\usage{
updateDefinition(definition, baseUrl, category, displayWarnings = TRUE)
}
\arguments{
\item{definition}{An R list object containing the expression for the specification. This will be
converted to JSON expression by function and posted into the WebApi.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{category}{These are the categories in WebApi. The valid string options are 'conceptSet', 
'cohort', 'characterization', 'pathway, 'incidenceRate', 'estimation', 'prediction'.}

\item{displayWarnings}{Display warnings returned by WebApi check}
}
\description{
Update definition \lifecycle{maturing}
}
\details{
Update a definition in WebAPI. Currently only cohorts are supported.
Takes the definition as a parameter and converts it to json. This is the full definition
(i.e. including name and id fields)
}
\examples{
\dontrun{
definition <- getDefinition(id = 13242, baseUrl = "http://server.org:80/WebAPI", category = "cohort")
definition$name <- "My new name for this"
updateDefinition(definition, baseUrl, category = "cohort")
}
}
