% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{updateEstimationDefinition}
\alias{updateEstimationDefinition}
\title{Update a Estimation definition. \lifecycle{maturing}}
\usage{
updateEstimationDefinition(
  estimationDefinition,
  baseUrl,
  displayWarnings = TRUE
)
}
\arguments{
\item{estimationDefinition}{An R list object containing the expression for the specification. Must
include id, name and expression. This will be converted to JSON
expression by function and posted into the WebApi. The definition will
be checked against the WebApi instance for errors}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{displayWarnings}{Display warnings returned by WebApi check}
}
\description{
Update a Estimation definition. \lifecycle{maturing}
}
\details{
Update a Estimation definition.
}
\examples{
\dontrun{
definition <- getEstimationDefinition(id = 13242,
                                      baseUrl = "http://server.org:80/WebAPI",
                                      category = estimation)
definition$name <- "My new name for this"
updateEstimation(estimationDefinition, baseUrl, category = "cohort")
}
}
