% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{updateIncidenceRateDefinition}
\alias{updateIncidenceRateDefinition}
\title{Update a IncidenceRate definition. \lifecycle{maturing}}
\usage{
updateIncidenceRateDefinition(
  incidenceRateDefinition,
  baseUrl,
  displayWarnings = TRUE
)
}
\arguments{
\item{incidenceRateDefinition}{An R list object containing the expression for the specification.
Must include id, name and expression. This will be converted to
JSON expression by function and posted into the WebApi. The
definition will be checked against the WebApi instance for errors}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{displayWarnings}{Display warnings returned by WebApi check}
}
\description{
Update a IncidenceRate definition. \lifecycle{maturing}
}
\details{
Update a IncidenceRate definition.
}
\examples{
\dontrun{
definition <- getIncidenceRateDefinition(id = 13242,
                                         baseUrl = "http://server.org:80/WebAPI",
                                         category = incidenceRate)
definition$name <- "My new name for this"
updateIncidenceRate(incidenceRateDefinition, baseUrl, category = "cohort")
}
}
