% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{updatePathwayDefinition}
\alias{updatePathwayDefinition}
\title{Update a Pathway definition. \lifecycle{maturing}}
\usage{
updatePathwayDefinition(pathwayDefinition, baseUrl, displayWarnings = TRUE)
}
\arguments{
\item{pathwayDefinition}{An R list object containing the expression for the specification. Must
include id, name and expression. This will be converted to JSON
expression by function and posted into the WebApi. The definition will be
checked against the WebApi instance for errors}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{displayWarnings}{Display warnings returned by WebApi check}
}
\description{
Update a Pathway definition. \lifecycle{maturing}
}
\details{
Update a Pathway definition.
}
\examples{
\dontrun{
definition <- getPathwayDefinition(id = 13242,
                                   baseUrl = "http://server.org:80/WebAPI",
                                   category = pathway)
definition$name <- "My new name for this"
updatePathway(pathwayDefinition, baseUrl, category = "cohort")
}
}
