% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{updatePredictionDefinition}
\alias{updatePredictionDefinition}
\title{Update a Prediction definition. \lifecycle{maturing}}
\usage{
updatePredictionDefinition(
  predictionDefinition,
  baseUrl,
  displayWarnings = TRUE
)
}
\arguments{
\item{predictionDefinition}{An R list object containing the expression for the specification. Must
include id, name and expression. This will be converted to JSON
expression by function and posted into the WebApi. The definition will
be checked against the WebApi instance for errors}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{displayWarnings}{Display warnings returned by WebApi check}
}
\description{
Update a Prediction definition. \lifecycle{maturing}
}
\details{
Update a Prediction definition.
}
\examples{
\dontrun{
definition <- getPredictionDefinition(id = 13242,
                                      baseUrl = "http://server.org:80/WebAPI",
                                      category = prediction)
definition$name <- "My new name for this"
updatePrediction(predictionDefinition, baseUrl, category = "cohort")
}
}
