% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auth.R
\name{authorizeWebApi}
\alias{authorizeWebApi}
\title{Authorize ROhdsiWebApi to access a protected instance of WebAPI Authorize the ROhdsiWebApi package
to access WebApi on behalf of the user. This can be done with any of the auth methods described
below. authorizeWebApi will use attempt to retrieve, cache, and update a token which will grant
access to webAPI by all subsequent requests made by the package.}
\usage{
authorizeWebApi(
  baseUrl,
  authMethod,
  webApiUsername = NULL,
  webApiPassword = NULL
)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{authMethod}{The method used for authentication to WebAPI. Options are
\itemize{
  \item {"db"}{Database authentication using Atlas/WebAPI built in auth}
  \item {"ad"}{Active Directory}
  \item {"windows"}{Windows NT authentication}
}
The auth method must be enabled in the instance of WebAPI pointed to by
baseUrl.}

\item{webApiUsername}{A character string containing the WebApi username passed on to
authentication methods}

\item{webApiPassword}{An character string containing a WebApi password passed on to authentication
methods. By default the user will be prompted for their password when
needed.}
}
\description{
Authorize ROhdsiWebApi to access a protected instance of WebAPI Authorize the ROhdsiWebApi package
to access WebApi on behalf of the user. This can be done with any of the auth methods described
below. authorizeWebApi will use attempt to retrieve, cache, and update a token which will grant
access to webAPI by all subsequent requests made by the package.
}
