% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{cancelCohortGeneration}
\alias{cancelCohortGeneration}
\title{Cancel generation of Cohort id.}
\usage{
cancelCohortGeneration(cohortId, baseUrl, sourceKey)
}
\arguments{
\item{cohortId}{An integer id representing the id that uniquely identifies a 
Cohort definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}
}
\value{
A tibble with job status information.
}
\description{
Cancel generation of Cohort id.
}
\details{
Cancel the generation of Cohort id in the WebApi.
}
\examples{
\dontrun{
cancelCohortGeneration(cohortId = 13242, 
baseUrl = "http://server.org:80/WebAPI",
sourceKey = 'HCUP')
}
}
