% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{convertConceptSetDefinitionToTable}
\alias{convertConceptSetDefinitionToTable}
\title{Convert a concept set definition to a table}
\usage{
convertConceptSetDefinitionToTable(conceptSetDefinition)
}
\arguments{
\item{conceptSetDefinition}{A concept set definition, for example as obtained through the \code{\link{getConceptSetDefinition}} function, 
or taken from a cohort definition.}
}
\value{
Takes a R (list) representation of the Concept Set expression and returns a table (dataframe)
representing the concept set expression. This is useful to create publication friendly output of
the concept set expression.
}
\description{
Convert a concept set definition to a table
}
\examples{
\dontrun{
conceptSetDefinition <- getConceptSetDefinition(conceptSetId = 282,
                                                baseUrl = "http://server.org:80/WebAPI")
convertConceptSetDefinitionToTable(conceptSetDefinition = conceptSetDefinition)
}

}
