% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{createConceptSetWorkbook}
\alias{createConceptSetWorkbook}
\title{Save a set of concept sets expressions, included concepts, and mapped concepts into a workbook}
\usage{
createConceptSetWorkbook(
  conceptSetIds,
  workFolder = NULL,
  baseUrl,
  included = FALSE,
  mapped = FALSE
)
}
\arguments{
\item{conceptSetIds}{A vector of concept set IDs.}

\item{workFolder}{Directory location where the workbook will be saved, defaults to working
directory.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{included}{Should included concepts be included in the workbook?}

\item{mapped}{Should mapped concepts be included in the workbook?}
}
\value{
A xlsx workbook (conceptSetExpressions.xlsx) that includes a list of all concept set IDs and names
and a worksheet for the concepts in each set. Options to include an included concepts and mapped
concepts worksheet for each concept set are avaialble.
}
\description{
Save a set of concept sets expressions, included concepts, and mapped concepts into a workbook
}
