% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\name{deleteCharacterizationDefinition}
\alias{deleteCharacterizationDefinition}
\title{Delete a cohort characterization definition and results}
\usage{
deleteCharacterizationDefinition(
  characterizationId,
  baseUrl,
  silent = FALSE,
  stopOnError = FALSE
)
}
\arguments{
\item{characterizationId}{The number indicating which characterization definition to fetch.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{silent}{[OPTIONAL, Default = FALSE] If TRUE, function will work silently without showing any warning or error message.}

\item{stopOnError}{[OPTIONAL, Default = FALSE] If silent silent = TRUE, then this will be ignored.}
}
\value{
NA. A status message will be shown.
}
\description{
Delete a cohort characterization definition and results
}
\details{
Deletes cohort characterization definition and results from WebAPI for a given characterization id
}
\examples{
\dontrun{
deleteCohortDefinition(characterizationId = 282, baseUrl = "http://server.org:80/WebAPI")
}

}
