% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{deleteCharacterizationDefinition}
\alias{deleteCharacterizationDefinition}
\title{Delete Characterization id definition. \lifecycle{stable}}
\usage{
deleteCharacterizationDefinition(characterizationId, baseUrl)
}
\arguments{
\item{characterizationId}{An integer id representing the id that uniquely identifies a
Characterization definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
None, unless error.
}
\description{
Delete Characterization id definition. \lifecycle{stable}
}
\details{
Delete the Characterization definition from WebAPI for a given Characterization id
}
\examples{
\dontrun{
deleteCharacterizationDefinition(characterizationId = 13242,
                                 baseUrl = "http://server.org:80/WebAPI")
}
}
