% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{deleteCohortDefinition}
\alias{deleteCohortDefinition}
\title{Delete Cohort id definition. \lifecycle{stable}}
\usage{
deleteCohortDefinition(cohortId, baseUrl)
}
\arguments{
\item{cohortId}{An integer id representing the id that uniquely identifies a Cohort definition in
a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
None, unless error.
}
\description{
Delete Cohort id definition. \lifecycle{stable}
}
\details{
Delete the Cohort definition from WebAPI for a given Cohort id
}
\examples{
\dontrun{
deleteCohortDefinition(cohortId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
