% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{deleteConceptSetDefinition}
\alias{deleteConceptSetDefinition}
\title{Delete ConceptSet id definition.}
\usage{
deleteConceptSetDefinition(conceptSetId, baseUrl)
}
\arguments{
\item{conceptSetId}{An integer id representing the id that uniquely identifies a ConceptSet
definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
None, unless error.
}
\description{
Delete ConceptSet id definition.
}
\details{
Delete the ConceptSet definition from WebAPI for a given ConceptSet id
}
\examples{
\dontrun{
deleteConceptSetDefinition(conceptSetId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
