% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{deleteEstimationDefinition}
\alias{deleteEstimationDefinition}
\title{Delete Estimation id definition.
\lifecycle{stable}}
\usage{
deleteEstimationDefinition(estimationId, baseUrl)
}
\arguments{
\item{estimationId}{An integer id representing the id that uniquely identifies a 
Estimation definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
None, unless error.
}
\description{
Delete Estimation id definition.
\lifecycle{stable}
}
\details{
Delete the Estimation definition from WebAPI for a given Estimation id
}
\examples{
\dontrun{
deleteEstimationDefinition(estimationId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
