% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{deletePathwayDefinition}
\alias{deletePathwayDefinition}
\title{Delete Pathway id definition.}
\usage{
deletePathwayDefinition(pathwayId, baseUrl)
}
\arguments{
\item{pathwayId}{An integer id representing the id that uniquely identifies a Pathway definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
None, unless error.
}
\description{
Delete Pathway id definition.
}
\details{
Delete the Pathway definition from WebAPI for a given Pathway id
}
\examples{
\dontrun{
deletePathwayDefinition(pathwayId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
