% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getCharacterizationDefinitionsMetaData}
\alias{getCharacterizationDefinitionsMetaData}
\title{Retrieve the meta data for Characterization definitions.}
\usage{
getCharacterizationDefinitionsMetaData(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A tibble of specification metadata for Characterization. Note: modifiedDate and createdDate are
returned as text/character.
}
\description{
Retrieve the meta data for Characterization definitions.
}
\details{
Obtains the meta data of WebApi specifications such as id, name, created/modified details, hash
object, etc. from WebApi for Characterization. This function is useful to retrieve the current
Characterization specifications.
}
\examples{
getCharacterizationDefinitionsMetaData(baseUrl = "http://server.org:80/WebAPI")

}
