% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vocabulary.R
\name{getClassificationFromSourceCode}
\alias{getClassificationFromSourceCode}
\title{Get classification term from source code}
\usage{
getClassificationFromSourceCode(baseUrl, sourceCode, sourceVocabularyId,
  targetVocabularyId, targetConceptClassId = NULL,
  relationshipId = "Has ancestor of", minDistance = 0,
  maxDistance = NULL)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{sourceCode}{A single source code}

\item{sourceVocabularyId}{The name of the source vocabulary id for the source code (e.g. ICD9CM, ICD10CM)}

\item{targetVocabularyId}{The classification vocabulary id to map to}

\item{targetConceptClassId}{(OPTIONAL) The concept class id of the classification term}

\item{relationshipId}{The name of the concept relationship to filter by. Refer to the concept_relationship table for a list of options.}

\item{minDistance}{The minimum hierarchy distance from the source code's standard concepts 
for the classification term. 
Default is 0 (meaning there is no minimum distance)}

\item{maxDistance}{The maximum hierarchy distance from the source code's standard concepts 
for the classification term. 
Default is NULL (meaning there is no maximum distance)}
}
\value{
A data frame summarizing the classification terms
}
\description{
Get classification term from source code
}
\details{
For a source code, obtain mapped classification term(s) by traversing from source code to
standard concept to classification term.
}
