% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\name{getCohortCharacterizationResults}
\alias{getCohortCharacterizationResults}
\title{Get Cohort Characterization Results}
\usage{
getCohortCharacterizationResults(
  characterizationId,
  baseUrl,
  generationId = NULL,
  sourceKey,
  cohortIds = c(),
  domains = c(),
  analysisNames = c()
)
}
\arguments{
\item{characterizationId}{The id of the cohort characterization in WebApi}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{generationId}{Used to specify the id of a particular generation of a cohort
characterization. By default, the latest execution is retrieved}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}

\item{cohortIds}{(OPTIONAL) Which cohort definition ids would you like to retrieve? By
default, all cohorts are retrieved.}

\item{domains}{(OPTIONAL) Which feature domains would you like to retrieve? By default,
all domains are retrieved.}

\item{analysisNames}{(OPTIONAL) Which feature analysis names would you like to retrieve? By
default, all analyses are retrieved.}
}
\description{
Get Cohort Characterization Results
}
