% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getCohortDefinition}
\alias{getCohortDefinition}
\title{Get Cohort id definition.}
\usage{
getCohortDefinition(cohortId, baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{CohortId}{An integer id representing the id that uniquely identifies a Cohort definition in
a WebApi instance.}
}
\value{
An R object representing the Cohort definition
}
\description{
Get Cohort id definition.
}
\details{
Obtain the Cohort definition from WebAPI for a given Cohort id
}
\examples{
\dontrun{
getCohortDefinition(CohortId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
