% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{getCohortDefinitionExpression}
\alias{getCohortDefinitionExpression}
\title{(Deprecated) Get a cohort definition expression}
\usage{
getCohortDefinitionExpression(cohortId, baseUrl)
}
\arguments{
\item{cohortId}{An integer id representing the id that uniquely identifies a cohort definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A JSON list object representing the cohort definition
This function has been deprecated. As an alternative please use the following
steps as in the example below:
  1) cohortDefinition <- getCohortDefinition(baseUrl = baseUrl, cohortId = 15873)
  2) validJsonExpression <- RJSONIO::toJSON(cohortDefinition$expression)
  3) save validJsonExpression object as .txt"
}
\description{
(Deprecated) Get a cohort definition expression
}
\details{
(Deprecated) Obtain the JSON expression from WebAPI for a given cohort id
}
\examples{
\dontrun{
# This will obtain a cohort definition's JSON expression:

getCohortDefinitionExpression(cohortId = 282, baseUrl = "http://server.org:80/WebAPI")
}

}
