% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{getCohortDefinitionName}
\alias{getCohortDefinitionName}
\title{(Deprecated) Get a cohort definition's name from WebAPI}
\usage{
getCohortDefinitionName(baseUrl, cohortId, formatName = FALSE)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{cohortId}{An integer id representing the id that uniquely identifies a cohort definition
in a WebApi instance.}

\item{formatName}{Should the name be formatted to remove prefixes and underscores?}
}
\value{
The name of the cohort.
}
\description{
(Deprecated) Get a cohort definition's name from WebAPI
}
\details{
(Deprecated) Obtains the name of a cohort. This function has been deprecated. As an alternative
please use getCohortDefinition
}
