% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getCohortDefinitionSql}
\alias{getCohortDefinitionSql}
\title{Get a cohort definition's SQL from WebAPI}
\usage{
getCohortDefinitionSql(cohortId, baseUrl, generateStats = TRUE)
}
\arguments{
\item{cohortId}{An integer id representing the id that uniquely identifies a cohort definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{generateStats}{Should the SQL include the code for generating inclusion rule statistics?
Note that if TRUE, several additional tables are expected to exists as
described in the details. By default this is TRUE.}
}
\value{
The templated SQL to generate the cohort
}
\description{
Get a cohort definition's SQL from WebAPI
}
\details{
Obtains the template SQL of a cohort. When using generateStats = TRUE, the following tables are
required to exist when executing the SQL: cohort_inclusion, cohort_inclusion_result,
cohort_inclusion_stats, and cohort_summary_stats. Also note that the cohort_inclusion table should
be populated with the names of the rules prior to executing the cohort definition SQL.
}
