% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getCohortGenerationInformation}
\alias{getCohortGenerationInformation}
\title{Get cohort generation information}
\usage{
getCohortGenerationInformation(cohortId, baseUrl, sourceKey)
}
\arguments{
\item{cohortId}{An integer id representing the id that uniquely identifies a cohort definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}
}
\value{
A data frame of cohort generation statuses, start times, and execution durations.
}
\description{
Get cohort generation information
}
\details{
Get information about cohort generation for a given combination of 
sourceKey and cohortId
}
