% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getCohortGenerationStatuses}
\alias{getCohortGenerationStatuses}
\title{Get Cohort Generation Statuses}
\usage{
getCohortGenerationStatuses(baseUrl, definitionIds, sourceKeys = NULL)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{definitionIds}{A list of cohort definition Ids}

\item{sourceKeys}{(OPTIONAL) A list of CDM source keys. These can be found in Atlas ->
Configure. Otherwise, all CDM source keys will be used.}
}
\value{
A data frame of cohort generation statuses, start times, and execution durations per definition id
and source key.
}
\description{
Get Cohort Generation Statuses
}
\details{
Obtains cohort generation statuses for a collection of cohort definition Ids and CDM sources.
Useful if running multiple cohort generation jobs that are long-running.
}
