% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getCohortInclusionRulesAndCounts}
\alias{getCohortInclusionRulesAndCounts}
\title{Get cohort inclusion rule counts for events and persons}
\usage{
getCohortInclusionRulesAndCounts(baseUrl, cohortId, sourceKey)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{cohortId}{The Atlas cohort definition id for the cohort}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the Configuration page}
}
\description{
Get cohort inclusion rule counts for events and persons
}
\details{
Obtains the inclusion rules from a cohort definition and summarizes the event and person counts per rule
}
