% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{getCohortInclusionRulesAndCounts}
\alias{getCohortInclusionRulesAndCounts}
\title{Get cohort inclusion rules and person counts}
\usage{
getCohortInclusionRulesAndCounts(baseUrl, cohortId, sourceKey)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{cohortId}{An integer id representing the id that uniquely identifies a cohort definition
in a WebApi instance.}

\item{sourceKey}{The source key for a CDM instance in WebAPI, as defined in the
Configuration page}
}
\description{
Get cohort inclusion rules and person counts
}
\details{
Obtains the inclusion rules from a cohort definition and summarizes the person counts per rule
}
