% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getCohortResults}
\alias{getCohortResults}
\title{Get results for a Cohort Id.}
\usage{
getCohortResults(cohortId, baseUrl)
}
\arguments{
\item{cohortId}{An integer id representing the id that uniquely identifies a cohort definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
An R object with results.
}
\description{
Get results for a Cohort Id.
}
\details{
Get the results for Cohort id.
}
\examples{
\dontrun{
getCohortResults(cohortId = 342, baseUrl = "http://server.org:80/WebAPI")
}
}
