% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getConceptSetConceptIds}
\alias{getConceptSetConceptIds}
\title{Get Concept Set Concept Ids}
\usage{
getConceptSetConceptIds(baseUrl, setId, vocabSourceKey = NULL)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{setId}{The concept set id in Atlas.}

\item{vocabSourceKey}{The source key of the Vocabulary. By default, the priority Vocabulary is
used.}
}
\value{
A list of concept Ids.
}
\description{
Get Concept Set Concept Ids
}
\details{
Obtains the full list of concept Ids in a concept set.
}
