% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getConceptSetDefinition}
\alias{getConceptSetDefinition}
\title{Get ConceptSet id definition. \lifecycle{stable}}
\usage{
getConceptSetDefinition(conceptSetId, baseUrl)
}
\arguments{
\item{conceptSetId}{An integer id representing the id that uniquely identifies a ConceptSet
definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
An R object representing the ConceptSet definition
}
\description{
Get ConceptSet id definition. \lifecycle{stable}
}
\details{
Obtain the ConceptSet definition from WebAPI for a given ConceptSet id
}
\examples{
\dontrun{
getConceptSetDefinition(conceptSetId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
