% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getConceptSetExpression}
\alias{getConceptSetExpression}
\title{Get a concept set expression}
\usage{
getConceptSetExpression(baseUrl, setId, asDataFrame = FALSE)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{setId}{The concept set id in WebApi}

\item{asDataFrame}{(OPTIONAL) Get expression as data frame}
}
\value{
A JSON list object representing the concept set
}
\description{
Get a concept set expression
}
\details{
Obtain the JSON expression from WebAPI for a given concept set
}
\examples{
\dontrun{
# This will obtain a concept set's JSON expression:

getConceptSetExpression(setId = 282, baseUrl = "http://server.org:80/WebAPI")
}

}
