% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{getConceptSetName}
\alias{getConceptSetName}
\title{Get a concept set's name from WebAPI}
\usage{
getConceptSetName(baseUrl, conceptSetId, formatName = FALSE)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{conceptSetId}{An integer id representing the id that uniquely identifies a 
concept set definition in a WebApi instance.}

\item{formatName}{Should the name be formatted to remove prefixes and underscores?}
}
\value{
The name of the concept set.
}
\description{
Get a concept set's name from WebAPI
}
\details{
Obtains the name of a concept set.
}
