% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{getEstimation}
\alias{getEstimation}
\title{Get an estimation definition}
\usage{
getEstimation(estimationId, baseUrl)
}
\arguments{
\item{estimationId}{The number indicating which estimation specification to fetch.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
An Robject representing the estimation defintion
}
\description{
Get an estimation definition
}
\details{
Obtain the estimation study specification from WebAPI for a given estimation id
}
\examples{
\dontrun{
estimation = getEstimation(estimationId = 124, 
                           baseUrl = "http://server.org:80/WebAPI")
}

}
