% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{getEstimationDefinition}
\alias{getEstimationDefinition}
\title{Get the estimation definition}
\usage{
getEstimationDefinition(baseUrl, estimationId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{estimationId}{The WebApi id for the estimation specification}
}
\value{
An R object.
}
\description{
Get the estimation definition
}
\details{
Get an R object representing the estimation definition.
}
\examples{
\dontrun{
getEstimationDefinition(baseUrl = "http://server.org:80/WebAPI", estimationId = 3434)
}

}
