% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getIncidenceRateDefinition}
\alias{getIncidenceRateDefinition}
\title{Get IncidenceRate id definition.}
\usage{
getIncidenceRateDefinition(incidenceRateId, baseUrl)
}
\arguments{
\item{incidenceRateId}{An integer id representing the id that uniquely identifies a IncidenceRate
definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
An R object representing the IncidenceRate definition
}
\description{
Get IncidenceRate id definition.
}
\details{
Obtain the IncidenceRate definition from WebAPI for a given IncidenceRate id
}
\examples{
\dontrun{
getIncidenceRateDefinition(incidenceRateId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
