% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getIncidenceRateDefinitionsMetaData}
\alias{getIncidenceRateDefinitionsMetaData}
\title{Retrieve the meta data for IncidenceRate definitions.}
\usage{
getIncidenceRateDefinitionsMetaData(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A tibble of specification metadata for IncidenceRate. Note: modifiedDate and createdDate are
returned as text/character.
}
\description{
Retrieve the meta data for IncidenceRate definitions.
}
\details{
Obtains the meta data of WebApi specifications such as id, name, created/modified details, hash
object, etc. from WebApi for IncidenceRate. This function is useful to retrieve the current
IncidenceRate specifications.
}
\examples{
getIncidenceRateDefinitionsMetaData(baseUrl = "http://server.org:80/WebAPI")

}
