% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceRate.R
\name{getIncidenceRateGenerationInformation}
\alias{getIncidenceRateGenerationInformation}
\title{Get incident rate generation information}
\usage{
getIncidenceRateGenerationInformation(incidenceRateId, baseUrl)
}
\arguments{
\item{incidenceRateId}{The WebApi ID for incidence rate analysis.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A tibble data-frame R-object with incidence rate
                       generation information.
}
\description{
Get incident rate generation information
}
\details{
Get incident rate generation information
}
\examples{
\dontrun{
getIncidenceRateGenerationInformation(incidenceRateId = 296, baseUrl = baseUrl)
                                         )
}
}
