% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R, R/IncidenceRate.R
\name{getIncidenceRateResults}
\alias{getIncidenceRateResults}
\title{Get results for a IncidenceRate Id.}
\usage{
getIncidenceRateResults(
  incidenceRateId,
  baseUrl,
  sourceKey,
  targetCohortId,
  outcomeCohortId
)

getIncidenceRateResults(
  incidenceRateId,
  baseUrl,
  sourceKey,
  targetCohortId,
  outcomeCohortId
)
}
\arguments{
\item{incidenceRateId}{The WebApi ID for incidence rate analysis.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{sourceKey}{The sourceKey for the incident rate analysis.}

\item{targetCohortId}{The cohort id of the target cohort in the incident rate analysis.}

\item{outcomeCohortId}{The cohort id of the outcome cohort in the incidence rate analysis.}
}
\value{
An R object with results.

A list with incident rate results summary, stratified results and treemap data.
}
\description{
Get results for a IncidenceRate Id.

Get results from incidence rate analysis.
}
\details{
Get the results for IncidenceRate id.

Given a sourceKey, targetCohortId, outcomeCohortId combinations, this function will return the
results of incidence rate analysis.
}
\examples{
\dontrun{
getIncidenceRateResults(incidenceRateId = 342, baseUrl = "http://server.org:80/WebAPI")
}
\dontrun{
getIncidenceRateResults(incidenceRateId = 296,
                        baseUrl = "http://server.org:80/WebAPI",
                        targetCohortId = 432423,
                        outcomeCohortId = 324,
                        sourceKey = "HCUP")
}
}
