% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getIncidenceRateResults}
\alias{getIncidenceRateResults}
\title{Get results for a IncidenceRate Id.}
\usage{
getIncidenceRateResults(incidenceRateId, baseUrl)
}
\arguments{
\item{incidenceRateId}{An integer id representing the id that uniquely identifies a incidence rate analysis definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
An R object with results.
}
\description{
Get results for a IncidenceRate Id.
}
\details{
Get the results for IncidenceRate id.
}
\examples{
\dontrun{
getIncidenceRateResults(incidenceRateId = 342, 
baseUrl = "http://server.org:80/WebAPI")
}
}
