% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getPathwayDefinition}
\alias{getPathwayDefinition}
\title{Get Pathway id definition.}
\usage{
getPathwayDefinition(pathwayId, baseUrl)
}
\arguments{
\item{pathwayId}{An integer id representing the id that uniquely identifies a Pathway definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
An R object representing the Pathway definition
}
\description{
Get Pathway id definition.
}
\details{
Obtain the Pathway definition from WebAPI for a given Pathway id
}
\examples{
\dontrun{
getPathwayDefinition(pathwayId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
