% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getPredictionDefinition}
\alias{getPredictionDefinition}
\title{Get Prediction id definition.
\lifecycle{stable}}
\usage{
getPredictionDefinition(predictionId, baseUrl)
}
\arguments{
\item{predictionId}{An integer id representing the id that uniquely identifies a 
Prediction definition in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
An R object representing the Prediction definition
}
\description{
Get Prediction id definition.
\lifecycle{stable}
}
\details{
Obtain the Prediction definition from WebAPI for a given Prediction id
}
\examples{
\dontrun{
getPredictionDefinition(predictionId = 13242, baseUrl = "http://server.org:80/WebAPI")
}
}
