% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{getPredictionDefinition}
\alias{getPredictionDefinition}
\title{Get a prediction definition}
\usage{
getPredictionDefinition(predictionId, baseUrl)
}
\arguments{
\item{predictionId}{An integer id representing the id that uniquely identifies a prediction definition
in a WebApi instance.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
An R object representing the cohort definition
}
\description{
Get a prediction definition
}
\details{
Obtain the prediction definition from WebAPI for a given prediction id
}
\examples{
\dontrun{
getPredictionDefinition(predictionId = 282, baseUrl = "http://server.org:80/WebAPI")
}

}
