% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{getPredictionDefinitionsMetaData}
\alias{getPredictionDefinitionsMetaData}
\title{Retrieve the meta data for Prediction definitions.}
\usage{
getPredictionDefinitionsMetaData(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A tibble of specification metadata for Prediction. Note: modifiedDate and createdDate are returned
as text/character.
}
\description{
Retrieve the meta data for Prediction definitions.
}
\details{
Obtains the meta data of WebApi specifications such as id, name, created/modified details, hash
object, etc. from WebApi for Prediction. This function is useful to retrieve the current Prediction
specifications.
}
\examples{
getPredictionDefinitionsMetaData(baseUrl = "http://server.org:80/WebAPI")

}
