% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{getUsersFromRole}
\alias{getUsersFromRole}
\title{Obtains users for a specific Atlas user role. Requires security enabled Atlas. \lifecycle{experimental}}
\usage{
getUsersFromRole(baseUrl, roleId)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{roleId}{The role id as defined in WebAPI sec_role table}
}
\value{
A data frame of user information for the Atlas role.
}
\description{
Obtains users for a specific Atlas user role. Requires security enabled Atlas. \lifecycle{experimental}
}
\details{
For a security enabled Atlas, obtains the user names for a specified Atlas user role. 
Role Id is set in WebAPI sec_role table
}
\examples{
\dontrun{
getUsersFromRole(baseUrl = "http://server.org:80/WebAPI",
                 roleId = 10)
}
}
