% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{getWebApiVersion}
\alias{getWebApiVersion}
\title{Get the version of the WebAPI}
\usage{
getWebApiVersion(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
The WebAPI version
}
\description{
Get the version of the WebAPI
}
\details{
Obtains the WebAPI version number
}
