% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{getWebApiVersion}
\alias{getWebApiVersion}
\title{Get the WebAPI version number \lifecycle{stable}}
\usage{
getWebApiVersion(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
The WebApi verions as a string.
}
\description{
Get the WebAPI version number \lifecycle{stable}
}
\details{
Obtains the WebAPI version number. This function is used to check that 
WebAPI baseUrl can be accessed and is a good first check to make sure 
you can access a WebAPI endpoint.
}
\examples{
\dontrun{
getWebApiVersion("http://server.org:80/WebAPI")
}
}
