% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebApi.R
\name{getWebApiVersion}
\alias{getWebApiVersion}
\title{Get the WebAPI version number \lifecycle{stable}}
\usage{
getWebApiVersion(baseUrl)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A string.
}
\description{
Get the WebAPI version number \lifecycle{stable}
}
\details{
Obtains the WebAPI version number.
}
