% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDefinition.R
\name{invokeCohortSetGeneration}
\alias{invokeCohortSetGeneration}
\title{Invoke the generation of a set of cohort definitions}
\usage{
invokeCohortSetGeneration(baseUrl, sourceKeys, cohortIds)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{sourceKeys}{A list of sourceKeys}

\item{cohortIds}{A list of cohortIds}
}
\description{
Invoke the generation of a set of cohort definitions
}
\details{
Invokes the generation of a set of cohort definitions across a set of CDMs set up in WebAPI. Use
\code{getCohortGenerationStatuses} to check the progress of the set.
}
