% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{isValidConceptSetId}
\alias{isValidConceptSetId}
\title{is ConceptSet id a valid definition in the WebApi.
\lifecycle{stable}}
\usage{
isValidConceptSetId(conceptSetIds, baseUrl)
}
\arguments{
\item{conceptSetIds}{A list of integer id(s) of the ConceptSet to be tested for validity.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
A logical vector indicating if an ID is valid.
}
\description{
is ConceptSet id a valid definition in the WebApi.
\lifecycle{stable}
}
\details{
Checks if a set of id for a ConceptSet is valid. The following checks are 
performed. 1) checks if all the ids exists in the WebApi i.e. valid.
}
\examples{
\dontrun{
isValidConceptSetId(conceptSetIds = c(13242, 3423, 34), baseUrl = "http://server.org:80/WebAPI")
}
}
