% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{isValidEstimationId}
\alias{isValidEstimationId}
\title{is Estimation id a valid definition in the WebApi. \lifecycle{stable}}
\usage{
isValidEstimationId(estimationIds, baseUrl)
}
\arguments{
\item{estimationIds}{A list of integer id(s) of the Estimation to be tested for validity.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A logical vector indicating if an ID is valid.
}
\description{
is Estimation id a valid definition in the WebApi. \lifecycle{stable}
}
\details{
Checks if a set of id for a Estimation is valid. The following checks are performed. 1) checks if
all the ids exists in the WebApi i.e. valid.
}
\examples{
\dontrun{
isValidEstimationId(estimationIds = c(13242, 3423, 34), baseUrl = "http://server.org:80/WebAPI")
}
}
