% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{isValidPathwayId}
\alias{isValidPathwayId}
\title{is Pathway id a valid definition in the WebApi. \lifecycle{stable}}
\usage{
isValidPathwayId(pathwayIds, baseUrl)
}
\arguments{
\item{pathwayIds}{A list of integer id(s) of the Pathway to be tested for validity.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}
}
\value{
A logical vector indicating if an ID is valid.
}
\description{
is Pathway id a valid definition in the WebApi. \lifecycle{stable}
}
\details{
Checks if a set of id for a Pathway is valid. The following checks are performed. 1) checks if all
the ids exists in the WebApi i.e. valid.
}
\examples{
\dontrun{
isValidPathwayId(pathwayIds = c(13242, 3423, 34), baseUrl = "http://server.org:80/WebAPI")
}
}
