% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoGeneratedDefinitions.R
\name{isValidPredictionId}
\alias{isValidPredictionId}
\title{is Prediction id a valid definition in the WebApi.
\lifecycle{stable}}
\usage{
isValidPredictionId(predictionIds, baseUrl)
}
\arguments{
\item{predictionIds}{A list of integer id(s) of the Prediction to be tested for validity.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}
}
\value{
A logical vector indicating if an ID is valid.
}
\description{
is Prediction id a valid definition in the WebApi.
\lifecycle{stable}
}
\details{
Checks if a set of id for a Prediction is valid. The following checks are 
performed. 1) checks if all the ids exists in the WebApi i.e. valid.
}
\examples{
\dontrun{
isValidPredictionId(predictionIds = c(13242, 3423, 34), baseUrl = "http://server.org:80/WebAPI")
}
}
