% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostDefinition.R
\name{postDefinition}
\alias{postDefinition}
\title{Post a definition into WebApi \lifecycle{maturing}}
\usage{
postDefinition(baseUrl, name, category, definition, duplicateNames)
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Note, there is no trailing '/'.
If trailing '/' is used, you may receive an error.}

\item{name}{A valid name for the definition. WebApi will use this name (if valid) as the
name of the definition. WebApi checks for validity, such as uniqueness, absence
of unacceptable character etc. An error might be thrown.}

\item{category}{These are the categories in WebApi. The valid string options are 'conceptSet', 
'cohort', 'characterization', 'pathway, 'incidenceRate', 'estimation', 'prediction'.}

\item{definition}{An R list object containing the expression for the specification. This will be
converted to JSON expression by function and posted into the WebApi. Note: only
limited checks are performed in R to check the validity of this expression.}

\item{duplicateNames}{How to handle importing a definition with a name that already exists in ATLAS.
'error' will throw an error, 
'overwrite' will attempt to overwrite the existing definition, 
'rename' will append the new defintion name with (1) until the name is unique}
}
\value{
This function will return a dataframe object with one row describing the posted WebApi expression
and its details. If unsuccessful a STOP message will be shown.
}
\description{
Post a definition into WebApi \lifecycle{maturing}
}
\details{
Post a definition into WebAPI. Currently only cohort and concept-set are supported.
}
\examples{
\dontrun{
definition <- getCohortDefinition(baseUrl = baseUrl, cohortId = 15873)
postDefinition(name = "new name for expression in sdaddaddd",
               baseUrl = "http://server.org:80/WebAPI",
               expression = definition,
               category = "cohort")
}
}
