% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSet.R
\name{resolveConceptSet}
\alias{resolveConceptSet}
\title{Resolve a concept set to the included standard concept IDs}
\usage{
resolveConceptSet(conceptSetDefinition, baseUrl, vocabularySourceKey = NULL)
}
\arguments{
\item{conceptSetDefinition}{A concept set definition, for example as obtained through the \code{\link{getConceptSetDefinition}} function, 
or taken from a cohort definition.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{vocabularySourceKey}{The source key of the Vocabulary. By default, the priority Vocabulary is
used.}
}
\value{
A vector of standard concept ids.
}
\description{
Resolve a concept set to the included standard concept IDs
}
\details{
Resolve a concept set to the included standard concept IDs
}
\examples{
\dontrun{
conceptSetDefinition <- getConceptSetDefinition(conceptSetId = 282,
                                                baseUrl = "http://server.org:80/WebAPI")
conceptIds <- resolveConceptSet(conceptSetDefinition = conceptSetDefinition,
                                baseUrl = "http://server.org:80/WebAPI")
}

}
