% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROhdsiWebApi.R
\docType{package}
\name{ROhdsiWebApi-package}
\alias{ROhdsiWebApi}
\alias{ROhdsiWebApi-package}
\title{ROhdsiWebApi: R package to allow interactivity with a WebApi instance}
\description{
Includes R wrappers for calling WebAPI web services.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/ROhdsiWebApi}
  \item Report bugs at \url{https://github.com/OHDSI/ROhdsiWebApi/issues}
}

}
\author{
\strong{Maintainer}: Ajit Londhe \email{alondhe@amgen.com}

Authors:
\itemize{
  \item Martijn Schuemie
  \item Frank DeFalco
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
