% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{setCohortPermission}
\alias{setCohortPermission}
\title{Assign cohort definition permissions to user. \lifecycle{experimental}}
\usage{
setCohortPermission(baseUrl, cohortId, userName, permission = "WRITE")
}
\arguments{
\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI".}

\item{cohortId}{The ID of the cohort definition}

\item{userName}{The userName to grant permission}

\item{permission}{The permission type to assign.  Can be "READ" or "WRITE". Defaults to "WRITE".}
}
\description{
Assign cohort definition permissions to user. \lifecycle{experimental}
}
\details{
The function will call the webservice to assign permissions to a cohort definition for the given userId.
}
\examples{
\dontrun{
setCohortPermission(baseUrl, cohortId, userName, permission = "WRITE") 
}
}
